% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/divergence_minimizers.R
\name{additive_mle}
\alias{additive_mle}
\title{Use mgcv to fit the projection to the posterior draws for additive multilevel
models.}
\usage{
additive_mle(formula, data, family, weights = NULL, ...)
}
\description{
Use mgcv to fit the projection to the posterior draws for additive multilevel
models.
}
