% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/augdat.R
\name{augdat_link_binom}
\alias{augdat_link_binom}
\title{Link function for augmented-data projection with binomial family}
\usage{
augdat_link_binom(prb_arr, link = "logit")
}
\arguments{
\item{prb_arr}{An array as described in section "Augmented-data projection"
of \code{\link[=extend_family]{extend_family()}}'s documentation.}

\item{link}{The same as argument \code{link} of \code{\link[=binomial]{binomial()}}.}
}
\value{
An array as described in section "Augmented-data projection" of
\code{\link[=extend_family]{extend_family()}}'s documentation.
}
\description{
This is the function which has to be supplied to \code{\link[=extend_family]{extend_family()}}'s argument
\code{augdat_link} in case of the augmented-data projection for the \code{\link[=binomial]{binomial()}}
family.
}
