% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formula.R
\name{collapse_contrasts_solution_path}
\alias{collapse_contrasts_solution_path}
\title{collapse a list of terms including contrasts}
\usage{
collapse_contrasts_solution_path(formula, path, data)
}
\arguments{
\item{formula}{model's formula}

\item{path}{list of terms possibly including contrasts}

\item{data}{model's data}
}
\value{
the updated list of terms replacing the contrasts with the term name
}
\description{
collapse a list of terms including contrasts
}
