% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formula.R
\name{count_terms_chosen}
\alias{count_terms_chosen}
\title{Given a refmodel structure, count the number of terms included.}
\usage{
count_terms_chosen(list_of_terms)
}
\arguments{
\item{list_of_terms}{Subset of terms from formula.}

\item{formula}{The reference model's formula.}
}
\description{
Given a refmodel structure, count the number of terms included.
}
