% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formula.R
\name{count_terms_in_formula}
\alias{count_terms_in_formula}
\title{Utility to count the number of terms in a given formula.}
\usage{
count_terms_in_formula(formula)
}
\arguments{
\item{subformula}{The right hand side of a formula for a valid model
either as a formula object or as a string.}
}
\value{
the number of terms in the subformula.
}
\description{
Utility to count the number of terms in a given formula.
}
