% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formula.R
\name{count_terms_in_group_term}
\alias{count_terms_in_group_term}
\title{Utility to count the number of terms in a given group term.}
\usage{
count_terms_in_group_term(term)
}
\arguments{
\item{term}{A group term as a string.}
}
\value{
the number of terms in the group term.
}
\description{
Utility to count the number of terms in a given group term.
}
