% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cv_fit.R
\name{cv_fit}
\alias{cv_fit}
\title{K-fold cross-validation for models fitted with rstanarm.}
\usage{
cv_fit(fit, k = 10)
}
\arguments{
\item{\code{fit}}{A \link[=stanreg-objects]{stanreg}-object.}

\item{\code{k}}{Number of cv-folds. Defaults to 10.}
}
\value{
An array with each row being one fold and first column containing
the fitted stan-model and the second column containing the corresponding testset.
}
\description{
Do a 10-fold cross-validation for a \link[=stanreg-objects]{stanreg}-object.
}
\examples{
\dontrun{
### Usage with stanreg objects
fit <- stan_glm(y~x, binomial())
fits <- cv_fit(fit)
# do cross-validated variable selection
vars <- cv_varsel(fit, fits)
}

}

