% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formula.R
\name{delete.intercept}
\alias{delete.intercept}
\title{remove intercept from formula}
\usage{
delete.intercept(formula)
}
\arguments{
\item{formula}{a model formula}
}
\value{
the updated formula without intercept
}
\description{
remove intercept from formula
}
