% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_funs.R
\name{.tabulate_stats}
\alias{.tabulate_stats}
\title{copied from summary_funs to remove duplicated code}
\usage{
.tabulate_stats(varsel, stats, alpha = 0.05, nfeat_baseline = NULL)
}
\description{
copied from summary_funs to remove duplicated code
}
