% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formula.R
\name{eval_rhs}
\alias{eval_rhs}
\title{Helper function to evaluate right hand side formulas in a context}
\usage{
eval_rhs(formula, data)
}
\arguments{
\item{formula}{Formula to evaluate.}

\item{data}{Data with which to evaluate.}
}
\value{
output from the evaluation
}
\description{
Helper function to evaluate right hand side formulas in a context
}
