% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/families.R
\name{extra-families}
\alias{extra-families}
\alias{Student_t}
\title{Extra family objects.}
\usage{
Student_t(link = "identity", nu = 3)
}
\arguments{
\item{link}{Specification of the link function, as for the default
\link[=family]{family}-objects.}

\item{nu}{Degrees of freedom for the Student-t distribution.}

\item{...}{Further arguments, currently ignored.}
}
\value{
A family object analogous to those described in
  \link[=family]{family}
}
\description{
Family objects not in the set of default \link[=family]{family}-objects.
}
