% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/latent.R
\name{extract_eta}
\alias{extract_eta}
\title{Extract the posterior draws of the latent predictor}
\usage{
extract_eta(fit, data)
}
\arguments{
\item{fit}{A reference model written}

\item{data}{The data on which the reference model was trained}

\item{ref_predfun}{The predictive function of the reference model}
}
\value{
An array of the posterior draws from the latent predictor
}
\description{
Extract the posterior draws of the latent predictor
}
