% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formula.R
\name{extract_response}
\alias{extract_response}
\title{At any point inside projpred, the response can be a single object or instead
it can represent multiple outputs. In this function we recover the response/s
as a character vector so we can index the dataframe.}
\usage{
extract_response(response)
}
\arguments{
\item{response}{The response as retrieved from the formula object.}
}
\value{
the response as a character vector.
}
\description{
At any point inside projpred, the response can be a single object or instead
it can represent multiple outputs. In this function we recover the response/s
as a character vector so we can index the dataframe.
}
