% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formula.R
\name{extract_terms_response}
\alias{extract_terms_response}
\title{This function splits the given formula in response (left) and predictors
(right).}
\usage{
extract_terms_response(formula)
}
\arguments{
\item{formula}{Formula object that specifies a model.}
}
\value{
a list containing the plain linear terms, interaction terms, group
  terms, response and a boolean global intercept indicating whether the
  intercept is included or not.
}
\description{
This function splits the given formula in response (left) and predictors
(right).
}
