% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/latent.R
\name{fit_latent}
\alias{fit_latent}
\title{Fit the latent reference model}
\usage{
fit_latent(
  fit,
  eta_post_draws,
  data,
  latent_family = brms::brmsfamily("gaussian"),
  latent_div_minimizer = projpred:::linear_mle,
  latent_proj_predfun = projpred:::linear_proj_predfun,
  dis_latent = rep(1, length(eta_post_draws)),
  extract_model_data = projpred:::.extract_latent_model_data,
  ref_predfun = projpred:::.latent_ref_predfun,
  cv_fun = projpred:::.latent_cvfun
)
}
\arguments{
\item{fit}{A reference model written}

\item{eta_post_draws}{The posterior predictive draws of the latent predictor from
the reference model}

\item{data}{The data on which the reference model was trained}

\item{latent_div_minimizer}{The divergence minimising formula to be used
on the latent predictor}

\item{latent_proj_predfun}{The projection predictive function}

\item{dis_latent}{An array of the dispertion parameter for the latent predictor}

\item{extract_model_data}{A function to extract data from the model}

\item{ref_predfun}{The predictive function of the reference model}

\item{cv_fun}{A function for cross-validation}
}
\value{
A latent predictor reference model
}
\description{
This function takes in a reference model, the posterior draws of its latent
predictor, and the data on which the reference model was trained to fit a
reference model in terms of the latent predictor using a Gaussianity 
assumption.
}
