% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formula.R
\name{flatten_additive_terms}
\alias{flatten_additive_terms}
\title{Remove duplicated additive terms.}
\usage{
flatten_additive_terms(terms_)
}
\arguments{
\item{terms}{A vector of additive terms as strings.}
}
\value{
a vector of unique linear interaction terms.
}
\description{
Remove duplicated additive terms.
}
