% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formula.R
\name{flatten_formula}
\alias{flatten_formula}
\title{By combining different submodels we may arrive to a formula with repeated
terms.
This function gets rid of duplicated terms_.}
\usage{
flatten_formula(formula)
}
\arguments{
\item{formula}{A formula specifying a model.}
}
\value{
a formula without duplicated structure.
}
\description{
By combining different submodels we may arrive to a formula with repeated
terms.
This function gets rid of duplicated terms_.
}
