% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formula.R
\name{flatten_group_terms}
\alias{flatten_group_terms}
\title{Unifies group terms removing any duplicates.}
\usage{
flatten_group_terms(terms_)
}
\arguments{
\item{terms}{A vector of linear group terms as strings.}
}
\value{
a vector of unique group terms.
}
\description{
Unifies group terms removing any duplicates.
}
