% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formula.R
\name{flatten_individual_terms}
\alias{flatten_individual_terms}
\title{Remove duplicated linear terms.}
\usage{
flatten_individual_terms(terms_)
}
\arguments{
\item{terms}{A vector of linear terms as strings.}
}
\value{
a vector of unique linear individual terms.
}
\description{
Remove duplicated linear terms.
}
