% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formula.R
\name{flatten_interaction_terms}
\alias{flatten_interaction_terms}
\title{Remove duplicated linear interaction terms.}
\usage{
flatten_interaction_terms(terms_)
}
\arguments{
\item{terms}{A vector of linear interaction terms as strings.}
}
\value{
a vector of unique linear interaction terms.
}
\description{
Remove duplicated linear interaction terms.
}
