% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formula.R
\name{formula2str}
\alias{formula2str}
\title{convert a formula to a character string}
\usage{
formula2str(formula, rm = c(0, 0), space = c("rm", "trim"))
}
\arguments{
\item{formula}{a model formula}

\item{rm}{a vector of to elements indicating how many characters
should be removed at the beginning and end of the string respectively}

\item{space}{how should whitespaces be treated?}
}
\value{
a single character string or NULL
}
\description{
convert a formula to a character string
}
