% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formula.R
\name{formula_contains_additive_terms}
\alias{formula_contains_additive_terms}
\title{Checks whether a formula contains additive terms or not.}
\usage{
formula_contains_additive_terms(formula)
}
\arguments{
\item{formula}{A formula for a valid model.}
}
\value{
TRUE if the formula contains additive terms, FALSE otherwise.
}
\description{
Checks whether a formula contains additive terms or not.
}
