% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/search.R
\name{fsel}
\alias{fsel}
\title{The forward selection algorithm}
\usage{
fsel(p_full, d_train, family_kl, intercept, nv, regul, coef_init, verbose)
}
\description{
p_full contains the parameters of the full model, d_train contains the data
and p_sub contains the parameters of the submodel. b0 is a guess for the
initial weight vector in the submodel. In addition, if a subset of the
samples is used for the variable selection, p_clust contains parameters
of the full model using only those samples.
}

