% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/refmodel.R
\name{get-refmodel}
\alias{get-refmodel}
\alias{get_refmodel}
\alias{get_refmodel.refmodel}
\alias{get_refmodel.vsel}
\alias{get_refmodel.default}
\alias{get_refmodel.stanreg}
\alias{init_refmodel}
\title{Get reference model structure}
\usage{
get_refmodel(object, ...)

\method{get_refmodel}{refmodel}(object, ...)

\method{get_refmodel}{vsel}(object, ...)

\method{get_refmodel}{default}(
  object,
  data,
  formula,
  ref_predfun,
  proj_predfun,
  div_minimizer,
  fetch_data,
  family = NULL,
  wobs = NULL,
  folds = NULL,
  cvfits = NULL,
  offset = NULL,
  cvfun = NULL,
  dis = NULL,
  ...
)

\method{get_refmodel}{stanreg}(
  object,
  data = NULL,
  ref_predfun = NULL,
  proj_predfun = NULL,
  div_minimizer = NULL,
  folds = NULL,
  ...
)

init_refmodel(
  object,
  data,
  formula,
  family,
  ref_predfun = NULL,
  div_minimizer = NULL,
  proj_predfun = NULL,
  folds = NULL,
  extract_model_data = NULL,
  cvfun = NULL,
  cvfits = NULL,
  dis = NULL,
  aug_data = FALSE,
  ...
)
}
\arguments{
\item{object}{Object on which the reference model is created. See possible
types below.}

\item{...}{Arguments passed to the methods.}

\item{data}{Data on which the reference model was fitted.}

\item{formula}{Reference model's lme4-like formula.}

\item{ref_predfun}{Prediction function for the linear predictor of the
reference model. See section "Details" below.}

\item{proj_predfun}{Prediction function for the linear predictor of the
projections.}

\item{div_minimizer}{Maximum likelihood estimator for the underlying
projection.}

\item{fetch_data}{Wrapper function for fetching the data without directly
accessing it. It should have a prototype fetch_data(data, data_points,
newdata = NULL), where data_points is a vector of data indices and newdata,
if not NULL, is a data frame with new data for testing.}

\item{family}{A family object that represents the observation model for the
reference model.}

\item{wobs}{A weights vector for the observations in the data. The default is
a vector of ones.}

\item{folds}{Only used for K-fold variable selection. It is a vector of fold
indices for each data point in data.}

\item{cvfits}{Only used for K-fold variable selection. A list of K-fold
fitted objects on which reference models are created.}

\item{offset}{A vector of offsets per observation to add to the linear
predictor.}

\item{cvfun}{Only used for K-fold variable selection. A function that, given
a folds vector, fits a reference model per fold and returns the fitted
object.}

\item{dis}{A dispersion vector for each observation.}

\item{extract_model_data}{A function with prototype
extract_model_data(object, newdata, wrhs, orhs), where object is a
reference model fit, newdata is either NULL or a data frame with new
observations, wrhs is a right hand side formula to recover the weights from
the data frame and orhs is a right hand side formula to recover the offset
from the data frame.}

\item{aug_data}{A single logical value indicating whether to use the
augmented-data approach for projection (\code{TRUE}) or not (\code{FALSE}).
For more information, see section "Details" below.}
}
\value{
An object of type \code{refmodel} (the same type as returned by
  \link{init_refmodel}) that can be passed to all the functions that take the
  reference fit as the first argument, such as \link{varsel},
  \link{cv_varsel}, \link{project}, \link[=proj-pred]{proj_predict} and
  \link[=proj-pred]{proj_linpred}.
}
\description{
Generic function that can be used to create and fetch the reference model
structure for all those objects that have this method. All these
implementations are wrappers to the \code{\link{init_refmodel}}-function so
the returned object has the same type.
}
\details{
As soon as possible, more information concerning the augmented-data
  projection will be provided here. For now,
  \href{https://github.com/stan-dev/projpred/issues/70}{this GitHub issue}
  provides some basic information. For the augmented-data projection,
  \code{ref_predfun} has to return an \emph{augmented-rows matrix}. For the
  definition of an augmented-rows matrix, let \eqn{N} denote the number of
  observations (in the original---i.e., non-augmented---dataset). Let
  \eqn{C_{\mbox{cat}}}{C_cat} denote the number of response categories. Let
  \eqn{C_{\mbox{lat}}}{C_lat} denote the number of latent response categories
  (which typically equals \eqn{C_{\mbox{cat}} - 1}{C_cat - 1}). Let \eqn{C}
  denote either \eqn{C_{\mbox{cat}}}{C_cat} or \eqn{C_{\mbox{lat}}}{C_lat},
  whichever is appropriate in the context where it is used (e.g., for
  \code{ref_predfun}'s output, \eqn{C = C_{\mbox{lat}}}{C = C_lat}). Then an
  augmented-rows matrix is a matrix with \eqn{N \cdot C}{N x C} rows in
  \eqn{C} blocks of \eqn{N} rows, i.e., with the \eqn{N} observations nested
  in the \eqn{C} (latent) response categories. For ordered response
  categories, the \eqn{C} (latent) response categories (i.e., the row blocks)
  have to be sorted in increasing order (according to the ordered response).
  The columns of an augmented-rows matrix have to correspond to the \eqn{S}
  posterior draws, just like for the non-augmented-data projection. An
  augmented-rows matrix is of class \code{"augmat"} (inheriting from classes
  \code{"matrix"} and \code{"array"}) and needs to have the value of \eqn{N}
  stored in an attribute called \code{"nobs_orig"}.
}
\examples{
\donttest{
if (requireNamespace('rstanarm', quietly=TRUE)) {
  ### Usage with stanreg objects
  dat <- data.frame(y = rnorm(100), x = rnorm(100))
  fit <- rstanarm::stan_glm(y ~ x, family = gaussian(), data = dat)
  ref <- get_refmodel(fit)
  print(class(ref))

  # variable selection, use the already constructed reference model
  vs <- varsel(ref)
  # this will first construct the reference model and then execute
  # exactly the same way as the previous command (the result is identical)
  vs <- varsel(fit)
}
}

}
