% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_dispersion.R
\name{get_dispersion}
\alias{get_dispersion}
\title{Dispersion parameter}
\usage{
get_dispersion(fit)
}
\arguments{
\item{fit}{A stanreg object.}
}
\description{
Get the parameter related to dispersion for some models from a stanreg object.
}

