% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clust.R
\name{get_p_clust}
\alias{get_p_clust}
\title{Function(s) to perform the clustering over the samples}
\usage{
get_p_clust(mu, dis, nc, wsample = rep(1, dim(mu)[2]), cl = NULL)
}
\description{
Function(s) to perform the clustering over the samples
}

