% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formula.R
\name{get_replace_response}
\alias{get_replace_response}
\title{Utility to just replace the response in the data frame}
\usage{
get_replace_response(formula, terms_, split_formula = FALSE)
}
\arguments{
\item{formula}{A formula for a valid model.}

\item{terms_}{A vector of terms to subset.}

\item{split_formula}{If TRUE breaks the response down into single response
  formulas.
Default FALSE. It only works if `y` represents a multi-output response.}

\item{y}{The response vector. Default NULL.}

\item{data}{The original data frame for the full formula.}
}
\value{
a function that replaces the response in the data with arguments
}
\description{
Utility to just replace the response in the data frame
}
