% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glm_proj_fsel.R
\name{glm_proj_fsel}
\alias{glm_proj_fsel}
\title{GLM projection to submodel using forward selection}
\usage{
glm_proj_fsel(x, b_p, w, dis_p, n_out, n_sel, d, family, mc.cores)
}
\arguments{
\item{x}{A model matrix.}

\item{b_p}{Sampled estimates of the coefficient of the full model.}

\item{w}{Observation weights.}

\item{dis_p}{dispersion parameter of the full model.}

\item{n_out}{Number of samples used in the projection.}

\item{n_sel}{Number of samples used in the variable selection.}

\item{d}{Maximum number of features in the projection (incl. intercept).}

\item{family}{A \code{\link{family}}-object.}

\item{mc.cores}{Number of cores used.}
}
\description{
A helper function that calls the forward selection
and the projection functions.
}

