% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glm_proj_helper.R
\name{glm_proj_helper}
\alias{glm_proj_helper}
\title{GLM projection to submodel using forward selection}
\usage{
glm_proj_helper(x, b_p, w, dis_p, family, args)
}
\arguments{
\item{\code{x}}{A model matrix.}

\item{\code{b_p}}{Sampled estimates of the coefficient of the full model.}

\item{\code{w}}{Observation weights.}

\item{\code{dis_p}}{dispersion parameter of the full model.}

\item{\code{family}}{A \code{\link{family}}-object.}

\item{\code{args}}{For possible additional arguments, see the
documentation of \code{\link{glm_proj}}.}
}
\description{
A helper function that calls the forward selection
and the projection functions. In order to use another
selection method, replace chosen <- fsel(...) with another
function.
}

