% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{init_refmodel}
\alias{init_refmodel}
\title{Initialize the reference model}
\usage{
init_refmodel(x, y, family, mu = NULL, dis = NULL, offset = NULL,
  wobs = NULL, wsample = NULL, intercept = TRUE, loglik = NULL)
}
\description{
Initialize the reference model
}
