% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IRLS.R
\name{IRLS}
\alias{IRLS}
\title{Iteratively reweighted least squared}
\usage{
IRLS(p_full, d_train, family_kl, intercept, regul, coef_init, eps = 1e-12,
  max_it = 300)
}
\description{
Does the projection from the full model to the submodel.
}
\details{
Written by hand because glm.fit in R is too slow.
- eps and max_it cannot be changed currently.
- assumes d_train$x has intercept as the first column.
}

