% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formula.R
\name{is_next_submodel_redundant}
\alias{is_next_submodel_redundant}
\title{Utility that checks if the next submodel is redundant with the current one.}
\usage{
is_next_submodel_redundant(current, new)
}
\arguments{
\item{current}{A list of terms included in the current submodel.}

\item{new}{The new term to add to the submodel.}

\item{formula}{The reference models' formula.}
}
\value{
TRUE if the new term results in a redundant model, FALSE otherwise.
}
\description{
Utility that checks if the next submodel is redundant with the current one.
}
