% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/family_functions.R
\name{kl_gauss}
\alias{kl_gauss}
\title{Helper functions for the NR-algorithm}
\usage{
kl_gauss(mu_p, x, mu_q, w, dis_p, dis_q)
}
\description{
Contains functions for calculating KL divergence and its first and second
derivatives for different families and link functions.
}
\details{
The NR-algorithm is general in the sense that in order to add a new model /
link function one simply needs to implement corresponding KL-divergence,
it's first and second derivatives + first and second derivatives of the
mu-parameter w.r.t eta.
}

