% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kl_helpers.R
\name{kl_helpers}
\alias{kl_helpers}
\title{Model-specific helper functions.}
\usage{
kl_helpers(fam)
}
\description{
\code{kl_helpers(fam)} returns a family object augmented with functions for
KL divergence, log predictive density and dispersion.
}
\details{
Naming convention described in fsel.R.
}

