% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formula.R
\name{lhs}
\alias{lhs}
\title{Extract left hand side of a formula as a formula itself by removing the right
hand side.}
\usage{
lhs(x)
}
\arguments{
\item{x}{Formula}
}
\value{
updated formula with an intercept in the right hand side.
}
\description{
Extract left hand side of a formula as a formula itself by removing the right
hand side.
}
