% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/divergence_minimizers.R
\name{linear_multilevel_mle}
\alias{linear_multilevel_mle}
\title{Use lmer to fit the projection to the posterior draws for multilevel models.}
\usage{
linear_multilevel_mle(formula, data, family, weights = NULL, regul = NULL, ...)
}
\description{
Use lmer to fit the projection to the posterior draws for multilevel models.
}
