% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formula.R
\name{make_formula}
\alias{make_formula}
\title{Subsets a formula by the given terms.}
\usage{
make_formula(terms_, formula = NULL)
}
\arguments{
\item{terms_}{A vector of terms to subset from the right hand side.}
}
\value{
A formula object with the collapsed terms.
}
\description{
Subsets a formula by the given terms.
}
