% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/search.R
\name{ordered_search_terms}
\alias{ordered_search_terms}
\title{Ordered search terms}
\usage{
ordered_search_terms(ordered_terms)
}
\arguments{
\item{ordered_terms}{A character vector of predictor terms that should be
selected in their given order.}
}
\value{
A character vector that may be used as input for argument
\code{search_terms} of \code{\link[=varsel]{varsel()}} or \code{\link[=cv_varsel]{cv_varsel()}}.
}
\description{
A helper function to construct the input for argument \code{search_terms} of
\code{\link[=varsel]{varsel()}} or \code{\link[=cv_varsel]{cv_varsel()}} if the predictor terms should be selected in a
given order.
}
\examples{
\dontshow{if (requireNamespace("rstanarm", quietly = TRUE)) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
# Data:
dat_gauss <- data.frame(y = df_gaussian$y, df_gaussian$x)

# The "stanreg" fit which will be used as the reference model (with small
# values for `chains` and `iter`, but only for technical reasons in this
# example; this is not recommended in general):
fit <- rstanarm::stan_glm(
  y ~ X1 + X2 + X3 + X4 + X5, family = gaussian(), data = dat_gauss,
  QR = TRUE, chains = 2, iter = 500, refresh = 0, seed = 9876
)

# We will select X3, X1, and X2 in this order:
search_terms_ordered <- ordered_search_terms(
  ordered_terms = c("X3", "X1", "X2")
)

# Run varsel() (here without cross-validation and with small values for
# `nterms_max`, `nclusters`, and `nclusters_pred`, but only for the sake of
# speed in this example; this is not recommended in general):
vs <- varsel(fit, nclusters = 5, nclusters_pred = 10,
             search_terms = search_terms_ordered, seed = 5555)
# Now see, for example, `?print.vsel`, `?plot.vsel`, `?suggest_size.vsel`,
# and `?ranking` for possible post-processing functions.
\dontshow{\}) # examplesIf}
}
\seealso{
\code{\link[=varsel]{varsel()}}, \code{\link[=cv_varsel]{cv_varsel()}}, \code{\link[=forced_search_terms]{forced_search_terms()}}
}
