% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formula.R
\name{parse_additive_terms}
\alias{parse_additive_terms}
\title{Parse additive terms from a list of individual terms. These may include
individual smoothers s(x), interaction smoothers s(x, z), smooth intercepts
s(factor, bs = "re"), group level independent smoothers s(x, by = g), or
shared smoothers for low dimensional factors s(x, g, bs = "fs"), interaction
group level smoothers with same shape for high dimensional factors t2(x, z,
g, bs = c(., ., "re")), interaction group level independent smoothers te(x,
z, by = g), population level interaction smoothers te(x, z, bs = c(., .)),
s(x, z).}
\usage{
parse_additive_terms(terms)
}
\arguments{
\item{terms}{list of terms to parse}
}
\value{
a vector of smooth terms
}
\description{
Parse additive terms from a list of individual terms. These may include
individual smoothers s(x), interaction smoothers s(x, z), smooth intercepts
s(factor, bs = "re"), group level independent smoothers s(x, by = g), or
shared smoothers for low dimensional factors s(x, g, bs = "fs"), interaction
group level smoothers with same shape for high dimensional factors t2(x, z,
g, bs = c(., ., "re")), interaction group level independent smoothers te(x,
z, by = g), population level interaction smoothers te(x, z, bs = c(., .)),
s(x, z).
}
