% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cv_varsel.R
\name{parse_args_cv_varsel}
\alias{parse_args_cv_varsel}
\title{Auxiliary function for parsing the input arguments for specific cv_varsel.
This is similar in spirit to parse_args_varsel, that is, to avoid the main
function to become too long and complicated to maintain.}
\usage{
parse_args_cv_varsel(
  refmodel,
  cv_method = NULL,
  K = NULL,
  nclusters = NULL,
  nclusters_pred = NULL
)
}
\arguments{
\item{refmodel}{Reference model as extracted by get_refmodel}

\item{cv_method}{The cross-validation method, either 'LOO' or 'kfold'.
Default is 'LOO'.}

\item{K}{Number of folds in the k-fold cross validation. Default is 5 for
genuine reference models and 10 for datafits (that is, for penalized
maximum likelihood estimation).}
}
\description{
Auxiliary function for parsing the input arguments for specific cv_varsel.
This is similar in spirit to parse_args_varsel, that is, to avoid the main
function to become too long and complicated to maintain.
}
