% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{plot.props}
\alias{plot.props}
\alias{plot.ranking}
\title{Plot ranking proportions from fold-wise predictor rankings}
\usage{
\method{plot}{props}(x, text_angle = NULL, ...)

\method{plot}{ranking}(x, ...)
}
\arguments{
\item{x}{For \code{\link[=plot.props]{plot.props()}}: an object of class \code{props} (returned by
\code{\link[=props]{props()}}, possibly with \code{cumulate = TRUE}). For \code{\link[=plot.ranking]{plot.ranking()}}: an
object of class \code{ranking} (returned by \code{\link[=ranking]{ranking()}}) that \code{\link[=props]{props()}} will be
applied to internally before then calling \code{\link[=plot.props]{plot.props()}}.}

\item{text_angle}{Passed to argument \code{angle} of \code{\link[ggplot2:element]{ggplot2::element_text()}} for
the y-axis tick labels. In case of long predictor names, \code{text_angle = 45}
might be helpful (for example).}

\item{...}{For \code{\link[=plot.ranking]{plot.ranking()}}: arguments passed to \code{\link[=props.ranking]{props.ranking()}} and
\code{\link[=plot.props]{plot.props()}}. For \code{\link[=plot.props]{plot.props()}}: currently ignored.}
}
\value{
A \pkg{ggplot2} plotting object (of class \code{gg} and \code{ggplot}).
}
\description{
Plots the ranking proportions (see \code{\link[=props]{props()}}) from the fold-wise predictor
rankings in a cross-validation with fold-wise searches. This is a
visualization of the \emph{transposed} matrix returned by \code{\link[=props]{props()}}. The
proportions printed as text inside of the colored tiles are rounded to whole
percentage points (the plotted proportions themselves are not rounded).
}
\examples{
# TODO

}
\author{
Idea and original code by Aki Vehtari. Slight modifications of the
original code by Frank Weber, Yann McLatchie, and Sölvi Rögnvaldsson. Final
implementation in \pkg{projpred} by Frank Weber.
}
