% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{plot}
\alias{plot}
\title{Plot summary statistics related to variable selection}
\usage{
plot(object, ...)
}
\arguments{
\item{object}{The object returned by \link[=varsel]{varsel} or
\link[=cv_varsel]{cv_varsel}.}

\item{...}{Currently ignored.}
}
\description{
Plot summary statistics related to variable selection
}
