% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.R
\name{predict.glmproj}
\alias{predict.glmproj}
\title{Predict}
\usage{
\method{predict}{glmproj}(object, x, d)
}
\arguments{
\item{\code{object}}{A glmproj object.}

\item{\code{x}}{A model matrix.}

\item{\code{d}}{The number of features to be used in the prediction.
If not provided, uses all features.}
}
\value{
\code{n*s} matrix of predictions using \code{d} features.
}
\description{
A predict method for glmproj objects.
}

