% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{print-vsel}
\alias{print-vsel}
\alias{print.vsel}
\alias{print.cvsel}
\title{Print methods for vsel/cvsel objects}
\usage{
\method{print}{vsel}(x, digits = 2, ...)

\method{print}{cvsel}(x, digits = 2, ...)
}
\arguments{
\item{x}{An object of class vsel/cvsel.}

\item{digits}{Number of decimal places to be reported (2 by default).}

\item{...}{Further arguments passed to \code{\link{varsel_stats}}.}
}
\value{
Returns invisibly the data frame produced by \code{\link{varsel_stats}}.
}
\description{
The \code{print} methods for vsel/cvsel objects created by \code{\link{varsel}}
or \code{\link{cv_varsel}}) rely on \code{\link{varsel_stats}} to display
a summary of the results of the projection predictive variable selection.
}
