% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{print-vselsummary}
\alias{print-vselsummary}
\alias{print.vselsummary}
\title{Print methods for summary objects}
\usage{
\method{print}{vselsummary}(x, digits = 1, ...)
}
\arguments{
\item{x}{An object of class vselsummary.}

\item{digits}{Number of decimal places to be reported (1 by default).}

\item{...}{Currently ignored.}
}
\value{
Returns invisibly the output produced by
\code{\link{summary.vsel}}.
}
\description{
The \code{print} methods for summary objects created by
\code{\link{summary}} to display a summary of the results of the
projection predictive variable selection.
}
