% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/project.R
\name{print.projection}
\alias{print.projection}
\title{Print information about \code{\link[=project]{project()}} output}
\usage{
\method{print}{projection}(x, ...)
}
\arguments{
\item{x}{An object of class \code{projection} (returned by \code{\link[=project]{project()}}, possibly
as elements of a \code{list}).}

\item{...}{Currently ignored.}
}
\value{
The input object \code{x} (invisible).
}
\description{
This is the \code{\link[=print]{print()}} method for objects of class \code{projection}. This method
mainly exists to avoid cluttering the console when printing such objects
accidentally.
}
