% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/refmodel.R
\name{print.refmodel}
\alias{print.refmodel}
\title{Print information about a reference model object}
\usage{
\method{print}{refmodel}(x, ...)
}
\arguments{
\item{x}{An object of class \code{refmodel} (returned by \code{\link[=get_refmodel]{get_refmodel()}} or
\code{\link[=init_refmodel]{init_refmodel()}}).}

\item{...}{Currently ignored.}
}
\value{
The input object \code{x} (invisible).
}
\description{
This is the \code{\link[=print]{print()}} method for reference model objects (objects of class
\code{refmodel}). This method mainly exists to avoid cluttering the console when
printing such objects accidentally.
}
