% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{print.vsel}
\alias{print.vsel}
\title{Print results (summary) of a \code{\link[=varsel]{varsel()}} or \code{\link[=cv_varsel]{cv_varsel()}} run}
\usage{
\method{print}{vsel}(x, digits = getOption("projpred.digits", 2), ...)
}
\arguments{
\item{x}{An object of class \code{vsel} (returned by \code{\link[=varsel]{varsel()}} or \code{\link[=cv_varsel]{cv_varsel()}}).}

\item{digits}{Passed to argument \code{digits} of \code{\link[=print.vselsummary]{print.vselsummary()}}.}

\item{...}{Arguments passed to \code{\link[=summary.vsel]{summary.vsel()}}.}
}
\value{
The output of \code{\link[=summary.vsel]{summary.vsel()}} (invisible).
}
\description{
This is the \code{\link[=print]{print()}} method for \code{vsel} objects (returned by \code{\link[=varsel]{varsel()}} or
\code{\link[=cv_varsel]{cv_varsel()}}). It displays a summary of a \code{\link[=varsel]{varsel()}} or \code{\link[=cv_varsel]{cv_varsel()}} run by
first calling \code{\link[=summary.vsel]{summary.vsel()}} and then \code{\link[=print.vselsummary]{print.vselsummary()}}.
}
