% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{proj-pred}
\alias{proj-pred}
\alias{proj_linpred}
\alias{proj_predict}
\title{Extract draws of the linear predictor and draw from the predictive
distribution of the projected submodel}
\usage{
proj_linpred(object, xnew, ynew = NULL, offsetnew = NULL,
  weightsnew = NULL, nv = NULL, transform = FALSE, integrated = FALSE,
  ...)

proj_predict(object, xnew, offsetnew = NULL, weightsnew = NULL, nv = NULL,
  draws = NULL, seed_samp = NULL, ...)
}
\arguments{
\item{object}{The object returned by \link{varsel},
\link{cv_varsel} or \link{project}.}

\item{xnew}{The predictor values used in the prediction. If \code{vind} is
specified, then the number and order of columns should correspond to
contain the variables at same positons (columns) as in the original data. If
\code{vind} is \code{NULL} and \code{xnew} is a matrix, it must contain the
same columns at same positions as in the original data. If \code{vind} is
\code{NULL} and \code{xnew} is a data frame, it must contain columns with
same names as in the original data.}

\item{ynew}{New (test) target variables.}

\item{offsetnew}{Offsets for the new observations. Defaults to a vector of
0s.}

\item{weightsnew}{Weights for the new observations. For binomial model,
corresponds to the number trials per observation. For \code{proj_linpred},
this argument matters only if \code{ynew} is specified. Defaults to a vector
of 1s.}

\item{nv}{Number of variables in the submodel (the variable combination is
taken from the \code{varsel} information). If a list, then results for all
specified model sizes are returned. Ignored if \code{vind} is specified.}

\item{transform}{Should the linear predictor be transformed using the
inverse-link function? Default is \code{FALSE}. \code{proj_linpred} only.}

\item{integrated}{If \code{TRUE}, the output is averaged over the
parameters. Default is \code{FALSE}. \code{proj_linpred} only.}

\item{...}{Additional argument passed to \link{project} if \code{object}
is an object returned by \link{varsel} or\link{cv_varsel}.}

\item{draws}{Number of draws to return from the predictive distribution of
the projection. The default is the number of draws in the projection.
\code{proj_predict} only.}

\item{seed_samp}{An optional seed to use for drawing from the projection.
\code{proj_predict} only.}
}
\value{
If nv has length one and ynew is unspecified, a matrix of
predictions. Otherwise, a list of length nv with the corresponding
predictions. If ynew is specified (proj_linpred only), a list with
elements pred (predictions) and lpd (corresponding log-densities)
for each submodel size.
}
\description{
\code{proj_linpred} extracts draws of the linear predictor and
\code{proj_predict} draws from the predictive distribution of the projected
submodel or submodels. If the projection has not been performed, the
functions also perform the projection.
}
