% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{proj_linpred}
\alias{proj_linpred}
\title{Linear predictor of a projected submodel}
\usage{
proj_linpred(object, xnew, ynew = NULL, offsetnew = NULL,
  weightsnew = NULL, transform = FALSE, integrated = FALSE, nv = NULL,
  vind = NULL, ns = NULL, nc = NULL)
}
\arguments{
\item{object}{The object returned by \link[=varsel]{varsel}, 
\link[=cv_varsel]{cv_varsel} or \link[=project]{project}.}

\item{xnew}{The predictor values used in the prediction. The number and order of the columns
should be the same as in the original full data if argument \code{nv} is specified (see below).
However, if argument \code{vind} is specified, then the number and order of columns should
correspond to \code{vind}.}

\item{ynew}{New (test) target variables.}

\item{offsetnew}{Offsets for the new observations.}

\item{weightsnew}{Weights for the new observations. This argument matters only if \code{ynew} is specified.}

\item{transform}{Should the linear predictor be transformed using the inverse-link function? 
Default is \code{FALSE}.}

\item{integrated}{If \code{TRUE}, the output is averaged over the
parameters. Default is \code{FALSE}.}

\item{nv}{Number of variables in the submodel (the variable combination is taken from the
\code{varsel} information). If a list, then results for all specified
model sizes are returned. Ignored if \code{vind} is specified.}

\item{vind}{Variable indices with which to predict. If specified, \code{nv} is ignored.}

\item{ns}{Number of samples to be projected. Ignored if \code{nc} is specified.}

\item{nc}{Number of clusters in the clustered projection. Default is 50.}

\item{intercept}{Whether to use intercept. Default is \code{TRUE}.}
}
\description{
Extract draws of the linear predictor from the projected submodel or 
submodels. If the projection has not been performed, the function
also performs the projection.
}
