% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proj_params.R
\name{proj_params}
\alias{proj_params}
\title{Project parameters}
\usage{
proj_params(mu_p, x, b_p, w, dis_p, funs, chosen, cores, verbose)
}
\arguments{
\item{\code{mu_p}}{Fitted values of the full model.}

\item{\code{x}}{A model matrix.}

\item{\code{b_p}}{Sampled estimates of the coefficient of the full model.}

\item{\code{w}}{Observation weights.}

\item{\code{dis_p}}{Dispersion parameter of the full model.}

\item{\code{funs}}{Model-specific helper functions.}

\item{\code{chosen}}{Indices for features used in the submodel.}

\item{\code{cores}}{Number of cores used.}

\item{\code{verbose}}{If \code{TRUE}, intermediate progress is printed (if only one core is used).}
}
\description{
Find parameters that minimize the KL divergence between the full model
and the submodel.
}

