% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{proj_pred}
\alias{proj_pred}
\alias{proj_linpred}
\alias{proj_predict}
\title{Extract draws of the linear predictor and draw from the predictive
distribution of the projected submodel}
\usage{
proj_linpred(object, xnew, ynew = NULL, offsetnew = NULL,
  weightsnew = NULL, transform = FALSE, integrated = FALSE, nv = NULL,
  vind = NULL, ns = NULL, nc = NULL, regul = 1e-09)

proj_predict(object, xnew, offsetnew = NULL, weightsnew = NULL, nv = NULL,
  vind = NULL, ns = NULL, nc = NULL, draws = NULL, seed = NULL,
  regul = 1e-09)
}
\arguments{
\item{object}{The object returned by \link[=varsel]{varsel},
\link[=cv_varsel]{cv_varsel} or \link[=project]{project}.}

\item{xnew}{The predictor values used in the prediction. The number and 
order of the columns should be the same as in the original full data if
argument \code{nv} is specified (see below). However, if argument 
\code{vind} is specified, then the number and order of columns should
correspond to \code{vind}.}

\item{ynew}{New (test) target variables.}

\item{offsetnew}{Offsets for the new observations. Defaults to a vector of
0s.}

\item{weightsnew}{Weights for the new observations. For binomial model,
corresponds to the number trials per observation. For \code{proj_linpred},
this argument matters only if \code{ynew} is specified. Defaults to a vector
of 1s.}

\item{transform}{Should the linear predictor be transformed using the
inverse-link function? Default is \code{FALSE}.}

\item{integrated}{If \code{TRUE}, the output is averaged over the
parameters. Default is \code{FALSE}.}

\item{nv}{Number of variables in the submodel (the variable combination is
taken from the \code{varsel} information). If a list, then results for all
specified model sizes are returned. Ignored if \code{vind} is specified.}

\item{vind}{Variable indices with which to predict. If specified, \code{nv}
is ignored.}

\item{ns}{Number of draws to be projected. Ignored if \code{nc} is specified.}

\item{nc}{Number of clusters in the clustered projection. Default is 50.}

\item{regul}{Amount of regularization in the projection. Usually there is no need for 
regularization, but sometimes for some models the projection can be ill-behaved and we
need to add some regularization to avoid numerical problems. Default is 1e-9.}

\item{draws}{Number of draws to return for \code{proj_predict}. Default is
\code{nc}.}

\item{intercept}{Whether to use intercept. Default is \code{TRUE}.}
}
\description{
\code{proj_linpred} extracts draws of the linear predictor and
\code{proj_predict} draws from the predictive distribution of the projected
submodel or submodels. If the projection has not been performed, the
functions also perform the projection.
}
