% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/projfun.R
\name{project_gaussian}
\alias{project_gaussian}
\title{Function handles for the projection}
\usage{
project_gaussian(ind, p_full, d_train, intercept, regul = 1e-12)
}
\description{
Function handles for the projection
}

