% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formula.R
\name{reduce_models}
\alias{reduce_models}
\title{Utility to remove redundant models to consider}
\usage{
reduce_models(chosen)
}
\arguments{
\item{chosen}{A list of included terms at a given point of the search.}

\item{refmodel}{The reference model's formula.}
}
\value{
a vector of incremental non redundant submodels for all the possible
  terms included.
}
\description{
Utility to remove redundant models to consider
}
