% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/search.R
\name{search_L1_surrogate}
\alias{search_L1_surrogate}
\title{copied over from search until we resolve the TODO below}
\usage{
search_L1_surrogate(
  p_ref,
  d_train,
  family,
  intercept,
  nterms_max,
  penalty,
  opt
)
}
\description{
copied over from search until we resolve the TODO below
}
