% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formula.R
\name{select_possible_terms_size}
\alias{select_possible_terms_size}
\title{Select next possible terms without surpassing a specific size}
\usage{
select_possible_terms_size(chosen, terms, size)
}
\arguments{
\item{chosen}{A list of currently chosen terms}

\item{terms}{A list of all possible terms}

\item{size}{Maximum allowed size}
}
\description{
Select next possible terms without surpassing a specific size
}
