% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formula.R
\name{sort_submodels_by_size}
\alias{sort_submodels_by_size}
\title{Given a list of formulas, sort them by the number of terms in them.}
\usage{
sort_submodels_by_size(submodels)
}
\arguments{
\item{submodels}{A list of models' formulas.}
}
\value{
a sorted list of submodels by included terms.
}
\description{
Given a list of formulas, sort them by the number of terms in them.
}
