% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formula.R
\name{split_additive_term}
\alias{split_additive_term}
\title{Plugs the main effects to the smooth additive term if `by` argument is
provided.}
\usage{
split_additive_term(term, data)
}
\arguments{
\item{term}{An additive term as a string.}

\item{data}{The reference model data.}
}
\value{
a minimally valid submodel for the additive term.
}
\description{
Plugs the main effects to the smooth additive term if `by` argument is
provided.
}
