% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formula.R
\name{split_formula}
\alias{split_formula}
\title{Simplify and split a formula by breaking it into all possible submodels.}
\usage{
split_formula(formula, return_group_terms = TRUE)
}
\arguments{
\item{formula}{A formula for a valid model.}

\item{return_group_terms}{If TRUE, return group terms as well. Default TRUE.}
}
\value{
a vector of all the minimal valid terms that make up for submodels.
}
\description{
Simplify and split a formula by breaking it into all possible submodels.
}
