% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formula.R
\name{split_group_term}
\alias{split_group_term}
\title{Simplify a single group term by breaking it down in as many terms_
as varying effects. It also explicitly adds or removes the varying intercept.}
\usage{
split_group_term(term)
}
\arguments{
\item{term}{A group term as a string.}
}
\value{
a vector of all the minimally valid submodels for the group term
including a single varying effect with and without varying intercept.
}
\description{
Simplify a single group term by breaking it down in as many terms_
as varying effects. It also explicitly adds or removes the varying intercept.
}
