% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formula.R
\name{split_interaction_term}
\alias{split_interaction_term}
\title{Plugs the main effects to the interaction terms to consider jointly for
projection.}
\usage{
split_interaction_term(term)
}
\arguments{
\item{term}{An interaction term as a string.}
}
\value{
a minimally valid submodel for the interaction term including
the overall effects.
}
\description{
Plugs the main effects to the interaction terms to consider jointly for
projection.
}
