% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_formula.R
\name{split_linear_term}
\alias{split_linear_term}
\title{Splits a linear term into individual predictors.}
\usage{
split_linear_term(term, data)
}
\arguments{
\item{term}{A matrix term.}

\item{data}{The original data frame.}
}
\value{
a list of the expanded linear matrix term.
}
\description{
Splits a linear term into individual predictors.
}
