% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formula.R
\name{subset_formula_and_data}
\alias{subset_formula_and_data}
\title{Utility to both subset the formula and update the data}
\usage{
subset_formula_and_data(formula, terms_, data, y = NULL, split_formula = FALSE)
}
\arguments{
\item{formula}{A formula for a valid model.}

\item{terms_}{A vector of terms to subset.}

\item{data}{The original data frame for the full formula.}

\item{y}{The response vector. Default NULL.}

\item{split_formula}{If TRUE breaks the response down into single response
  formulas.
Default FALSE. It only works if `y` represents a multi-output response.}
}
\value{
a list including the updated formula and data
}
\description{
Utility to both subset the formula and update the data
}
