% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{suggest_size}
\alias{suggest_size}
\alias{suggest_size.vsel}
\title{Suggest model size}
\usage{
suggest_size(object, ...)

\method{suggest_size}{vsel}(
  object,
  stat = "elpd",
  alpha = 0.32,
  pct = 0,
  type = "upper",
  baseline = NULL,
  warnings = TRUE,
  ...
)
}
\arguments{
\item{object}{An object of class \code{vsel} (returned by \code{\link[=varsel]{varsel()}} or
\code{\link[=cv_varsel]{cv_varsel()}}).}

\item{...}{Currently ignored.}

\item{stat}{Statistic used for the decision. See \code{\link[=summary.vsel]{summary.vsel()}} for
possible choices.}

\item{alpha}{A number giving the desired coverage of the credible intervals
based on which the decision is made. For example, \code{alpha = 0.32}
corresponds to 68\% probability mass within the intervals, that is,
one-standard-error intervals. See section "Details" for more information.}

\item{pct}{A number giving the relative proportion (\emph{not} percents) between
baseline model and null model utilities one is willing to sacrifice. See
section "Details" for more information.}

\item{type}{Either \code{"upper"} or \code{"lower"} determining whether the decisions
are based on the upper or lower credible bounds, respectively. See section
"Details" for more information.}

\item{baseline}{Either \code{"ref"} or \code{"best"} indicating whether the baseline is
the reference model or the best submodel found, respectively. If \code{NULL},
then \code{"ref"} is used, except for \code{datafit}s for which \code{"best"} is used.}

\item{warnings}{Whether to give warnings if automatic suggestion fails,
mainly for internal use. Usually there is no reason to set this to \code{FALSE}.}
}
\description{
This function can suggest an appropriate model size based on a decision rule
described in section "Details". Note that the decision rule is heuristic and
should only be interpreted as a guideline. It is recommended to examine the
results via \code{\link[=plot.vsel]{plot.vsel()}} and/or \code{\link[=summary.vsel]{summary.vsel()}} and make the final decision
based on what is most appropriate for the given problem.
}
\details{
The suggested model size is the smallest model size for which either
the lower or upper bound (depending on argument \code{type}) of the credible
interval (with coverage \code{1 - alpha}) for \eqn{u_k - u_{\mbox{base}}}{u_k -
  u_base} (with \eqn{u_k} denoting the \eqn{k}-th submodel's utility and
\eqn{u_{\mbox{base}}}{u_base} denoting the baseline model's utility) falls
above (or is equal to) \deqn{\mbox{pct} * (u_0 - u_{\mbox{base}})}{pct *
  (u_0 - u_base)} where \eqn{u_0} denotes the null model utility. The
baseline is either the reference model or the best submodel found (see
argument \code{baseline}).

For example, \code{alpha = 0.32}, \code{pct = 0}, and \code{type = "upper"} means that we
select the smallest model size for which the upper bound of the
corresponding credible interval exceeds (or is equal to) the baseline model
utility, that is, which is better than the baseline model with a
probability of at least 0.16 (and consequently, worse with a probability of
at most 0.84). In other words, the estimated difference between the
baseline model utility and the submodel utility is at most one standard
error away from zero, so the two utilities are considered to be close.
}
\note{
Loss statistics like the root mean-squared error (RMSE) and the
mean-squared error (MSE) are converted to utilities by multiplying them by
\code{-1}, so a call such as \code{suggest_size(object, stat = "rmse", type = "upper")} finds the smallest model size whose upper credible bound of the
\emph{negative} RMSE or MSE exceeds the cutoff (or equivalently has the lower
credible bound of RMSE or MSE below the cutoff). This is done to make the
interpretation of argument \code{type} the same regardless of argument \code{stat}.

The intercept is not counted by \code{\link[=suggest_size]{suggest_size()}}, so a suggested size of
zero stands for the intercept-only model.
}
\examples{
if (requireNamespace("rstanarm", quietly = TRUE)) {
  # Data:
  dat_gauss <- data.frame(y = df_gaussian$y, df_gaussian$x)

  # The "stanreg" fit which will be used as the reference model:
  fit <- rstanarm::stan_glm(
    y ~ X1 + X2 + X3 + X4 + X5, family = gaussian(), data = dat_gauss,
    QR = TRUE, chains = 2, iter = 500, refresh = 0, seed = 9876
  )

  # Variable selection (here without cross-validation and with small values
  # for `nterms_max`, `nclusters`, and `nclusters_pred`, but only for the
  # sake of speed in this example; this is not recommended in general):
  vs <- varsel(fit, nterms_max = 3, nclusters = 5, nclusters_pred = 10,
               seed = 5555)
  print(suggest_size(vs))
}

}
