% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formula.R
\name{to_character_rhs}
\alias{to_character_rhs}
\title{Cast a right hand side formula to a character vector.}
\usage{
to_character_rhs(rhs)
}
\arguments{
\item{rhs}{a right hand side formula of the type . ~ x + z + ...}
}
\value{
a character vector containing only the right hand side.
}
\description{
Cast a right hand side formula to a character vector.
}
