% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formula.R
\name{validate_resp_formula}
\alias{validate_resp_formula}
\title{validate formulas dedicated to response variables}
\usage{
validate_resp_formula(x, empty_ok = TRUE)
}
\arguments{
\item{x}{coerced to a formula object}

\item{empty_ok}{is an empty left-hand-side ok?}
}
\value{
a formula of the form <response> ~ 1
}
\description{
validate formulas dedicated to response variables
}
