% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formula.R
\name{validate_response_formula}
\alias{validate_response_formula}
\title{Because the formula can imply multiple or single response, in this function
we make sure that a formula has a single response. Then, in the case of
multiple responses we split the formula.}
\usage{
validate_response_formula(formula)
}
\arguments{
\item{formula}{A formula specifying a model.}
}
\value{
a formula or a list of formulas with a single response each.
}
\description{
Because the formula can imply multiple or single response, in this function
we make sure that a formula has a single response. Then, in the case of
multiple responses we split the formula.
}
