% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{varsel-methods}
\alias{varsel-methods}
\alias{proj_linpred}
\alias{proj_coef}
\alias{proj_se}
\alias{proj_sigma}
\title{Methods for objects with varsel-information}
\usage{
proj_linpred(object, xnew, ynew = NULL, offsetnew = NULL,
  transform = FALSE, nv = NULL, integrated = FALSE, ns = NULL,
  nc = NULL, manual_inds = NULL, ...)

proj_coef(object, ...)

proj_se(object, ...)

proj_sigma(object, ...)
}
\arguments{
\item{object}{The object returned by \link[=varsel]{varsel} or
\link[=cv_varsel]{cv_varsel}.}

\item{xnew}{New data to be used with the prediction. If \code{NULL}, the
data used when fitting the model is used.}

\item{ynew}{Same as \code{xnew}, but for the ouput.}

\item{offsetnew}{Same as \code{xnew}, but for the offset.}

\item{transform}{Same as in \link[=posterior_linpred]{posterior_linpred}.}

\item{nv}{Number of variables in the submodel. Can also be a list, in
which case a list is returned for each submodel size.}

\item{integrated}{If \code{TRUE}, the output is averaged over the
parameters. Defaults to \code{FALSE}.}

\item{ns}{Same as in \link[=varsel]{varsel}.}

\item{nc}{Same as in \link[=varsel]{varsel}.}

\item{...}{Currently ignored.}
}
\description{
The methods can be used to extract all the information related to variable
selection obtained running either \link[=varsel]{varsel} or
\link[=cv_varsel]{cv_varsel}. Most of the functions work similar to the
generic methods defined for objects of the class 'lm', 'glm', 'stanreg',
eg. proj_coef corresponds to \link[=coef]{coef}, proj_linpred corresponds to
\link[=se]{se} etc. Note that these are under development and subject might
change in the future.
}
