% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workflow.R
\name{workflow}
\alias{workflow}
\alias{varsel_search}
\alias{varsel_search.default}
\alias{varsel_search.refmodel}
\alias{approximate_loo}
\alias{approximate_loo.default}
\alias{approximate_loo.refmodel}
\alias{approximate_loo.vselsearch}
\alias{approximate_kfold}
\alias{approximate_kfold.default}
\alias{approximate_kfold.refmodel}
\alias{approximate_kfold.vselsearch}
\alias{cv_loo}
\alias{cv_loo.default}
\alias{cv_loo.refmodel}
\alias{cv_loo.vselsearch}
\alias{cv_loo.vselapproxcv}
\alias{cv_kfold}
\alias{cv_kfold.default}
\alias{cv_kfold.refmodel}
\alias{cv_kfold.vselsearch}
\alias{varsel_cv}
\alias{varsel_cv.default}
\alias{varsel_cv.refmodel}
\alias{varsel_cv.vselsearch}
\alias{varsel_cv.vselapproxcv}
\title{Variable and structure selection workflow}
\usage{
varsel_search(object, ...)

\method{varsel_search}{default}(object, ...)

\method{varsel_search}{refmodel}(
  object,
  method = NULL,
  ndraws = NULL,
  nclusters = NULL,
  nterms_max = NULL,
  verbose = TRUE,
  lambda_min_ratio = 1e-05,
  nlambda = 150,
  thresh = 1e-06,
  regul = 1e-04,
  penalty = NULL,
  search_terms = NULL,
  ...
)

approximate_loo(object, ...)

\method{approximate_loo}{default}(object, ...)

\method{approximate_loo}{refmodel}(object, ...)

\method{approximate_loo}{vselsearch}(
  object,
  ndraws_pred = NULL,
  nclusters_pred = NULL,
  verbose = TRUE,
  penalty = NULL,
  nloo = NULL,
  refit_proj = TRUE,
  seed = NULL,
  ...
)

approximate_kfold(object, ...)

\method{approximate_kfold}{default}(object, ...)

\method{approximate_kfold}{refmodel}(object, ...)

\method{approximate_kfold}{vselsearch}(
  object,
  ndraws_pred = NULL,
  nclusters_pred = NULL,
  K = NULL,
  verbose = TRUE,
  penalty = NULL,
  refit_proj = TRUE,
  seed = NULL,
  ...
)

cv_loo(object, ...)

\method{cv_loo}{default}(object, ...)

\method{cv_loo}{refmodel}(object, ...)

\method{cv_loo}{vselsearch}(object, ...)

\method{cv_loo}{vselapproxcv}(
  object,
  method = NULL,
  ndraws = NULL,
  nclusters = NULL,
  ndraws_pred = NULL,
  nclusters_pred = NULL,
  nterms_max = NULL,
  penalty = NULL,
  verbose = TRUE,
  nloo = NULL,
  lambda_min_ratio = 1e-05,
  nlambda = 150,
  thresh = 1e-06,
  regul = 1e-04,
  seed = NULL,
  search_terms = NULL,
  ...
)

cv_kfold(object, ...)

\method{cv_kfold}{default}(object, ...)

\method{cv_kfold}{refmodel}(object, ...)

\method{cv_kfold}{vselsearch}(object, ...)

varsel_cv(object, ...)

\method{varsel_cv}{default}(object, ...)

\method{varsel_cv}{refmodel}(object, ...)

\method{varsel_cv}{vselsearch}(object, ...)

\method{varsel_cv}{vselapproxcv}(
  object,
  method = NULL,
  cv_method = NULL,
  ndraws = NULL,
  nclusters = NULL,
  ndraws_pred = NULL,
  nclusters_pred = NULL,
  nterms_max = NULL,
  K = NULL,
  penalty = NULL,
  verbose = TRUE,
  nloo = NULL,
  lambda_min_ratio = 1e-05,
  nlambda = 150,
  thresh = 1e-06,
  regul = 1e-04,
  seed = NULL,
  search_terms = NULL,
  cv_search = TRUE,
  ...
)
}
\description{
Perform variable and structure selection in a cascade workflow. First, it
searches through model space to determine the ordering of model terms for
projection. Second, we compute the projections with approximate cross
validation. Third, if selection diagnostics indicate that more posterior
draws are required, we recommend running a full search with more draws.
}
