% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cv_varsel.R
\name{cv_varsel}
\alias{cv_varsel}
\title{Cross-validate the variable selection (varsel)}
\usage{
cv_varsel(fit, method = "L1", cv_method = "LOO", ns = NULL, nc = NULL,
  nspred = NULL, ncpred = NULL, nv_max = NULL, intercept = NULL,
  verbose = T, K = NULL, k_fold = NULL, ...)
}
\arguments{
\item{fit}{Same as in \link[=varsel]{varsel}.}

\item{method}{Same as in \link[=varsel]{varsel}.}

\item{cv_method}{The cross-validation method, either 'LOO' or 'kfold'. Default is 'LOO'.}

\item{ns}{Number of samples used for selection. Ignored if nc is provided or if method='L1'.}

\item{nc}{Number of clusters used for selection. Default is 1 and ignored if method='L1' 
(L1-search uses always one cluster).}

\item{nspred}{Number of samples used for prediction (after selection). Ignored if ncpred is given.}

\item{ncpred}{Number of clusters used for prediction (after selection). Default is 1.}

\item{nv_max}{Same as in \link[=varsel]{varsel}.}

\item{intercept}{Same as in \link[=varsel]{varsel}.}

\item{verbose}{Whether to print out some information during the validation, Default is TRUE.}

\item{K}{Number of folds in the k-fold cross validation. Only applicable
if \code{cv_method = TRUE} and \code{k_fold = NULL}.}

\item{k_fold}{An array with cross-validated stanfits and the respective
test datasets returned by \link[=kfold]{kfold} with \code{save_fits=TRUE}.
If not provided, \link[=kfold]{kfold} is called inside the function.}
}
\value{
The original \link[=stanreg-objects]{stanreg} object augmented with an element 'varsel',
which is a list containing the following elements:
\describe{
 \item{\code{chosen}}{The order in which the variables were added to the submodel.}
 \item{\code{pctch}}{Percentage of cross-validation runs that included the given
   variable to a model of given size.}
 \item{\code{kl}}{KL-divergence for each submodel size.}
 \item{\code{summaries}}{CV-summary statistics computed during the selection.}
 \item{\code{d_test}}{The data used to evaluate the summaries.}
 \item{\code{family_kl}}{A modified \code{\link{family}}-object.}
}
}
\description{
Perform cross-validation for the projective variable selection for a generalized
linear model.
}
\examples{
\dontrun{
### Usage with stanreg objects
fit <- stan_glm(y~x, binomial())
fit_cv <- cv_varsel(fit)
}

}
