% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{cvind}
\alias{cvind}
\title{Create cross-validation indices}
\usage{
cvind(n, k, seed = NULL)
}
\arguments{
\item{n}{Number of data points.}

\item{k}{Number of folds.}

\item{seed}{Random seed so that the same division could be obtained again if needed.}
}
\value{
A list with \code{k} elements, each having fields \code{itr} and \code{its} 
which give the training and test indices, respectively, for each fold.
}
\description{
Divide indices from 1 to \code{n} into subsets for \code{k}-fold cross validation.
This function is potentially useful for creating the cross-validation objects for 
\link[=init_refmodel]{init_refmodel}.
}
\examples{
\donttest{
### compute sample means within each fold
n <- 100
y <- rnorm(n)
cv <- cvind(n, k=5)
cvmeans <- lapply(cv, function(fold) mean(y[fold$itr]))
}
}
