% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{df_binom}
\alias{df_binom}
\title{Binomial toy example.}
\format{A simulated classification dataset containing 100 observations.
\describe{
  \item{y}{target, 0 or 1.}
  \item{x}{features, 30 in total.}
}}
\source{
\url{http://web.stanford.edu/~hastie/glmnet/glmnetData/BNExample.RData}
}
\usage{
df_binom
}
\description{
Binomial toy example.
}
\keyword{datasets}
