% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{df_gaussian}
\alias{df_gaussian}
\title{Gaussian toy example}
\format{
A simulated regression dataset containing 100 observations.
\describe{
\item{y}{response, real-valued.}
\item{x}{predictors, 20 in total. Mean and SD approximately 0 and 1,
respectively.}
}
}
\source{
\url{https://web.stanford.edu/~hastie/glmnet/glmnetData/QSExample.RData}
}
\usage{
df_gaussian
}
\description{
Gaussian toy example
}
\keyword{datasets}
