% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extend_family.R
\name{extend_family}
\alias{extend_family}
\title{Extend a family}
\usage{
extend_family(family)
}
\arguments{
\item{family}{A \code{\link{family}} object.}
}
\value{
The \code{\link{family}} object extended in the way needed by \pkg{projpred}.
}
\description{
This function adds some internally required elements to a \code{\link{family}} object.
It is called internally by \code{\link[=init_refmodel]{init_refmodel()}}, so you will rarely need to call
it yourself.
}
