% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extend_family.R
\name{extend_family}
\alias{extend_family}
\title{Add extra fields to the family object.}
\usage{
extend_family(family)
}
\arguments{
\item{family}{Family object.}
}
\value{
Extended family object.
}
\description{
Add extra fields to the family object.
}
