% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fsel.R
\name{fsel}
\alias{fsel}
\title{Forward selection}
\usage{
fsel(mu_p, x, b_p, w, dis_p, funs, avg, d, cores, verbose)
}
\arguments{
\item{\code{mu_p}}{Fitted values of the full model.}

\item{\code{x}}{A model matrix.}

\item{\code{b_p}}{Sampled estimates of the coefficient of the full model.}

\item{\code{w}}{Observation weights.}

\item{\code{dis_p}}{dispersion parameter of the full model.}

\item{\code{funs}}{Model-specific helper functions.}

\item{\code{avg}}{If TRUE, KL divergence of the average of the samples
is used instead of the average of the KL divergences for each sample.}

\item{\code{d}}{Maximum number of features in the projection.}

\item{\code{cores}}{Number of cores used.}

\item{\code{verbose}}{If \code{TRUE}, intermediate progress is printed (if only one core is used).}
}
\description{
Performs forward selection on the data and returns
the 'optimal' order of the parameters.
}

