% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fsel.R
\name{fsel}
\alias{fsel}
\title{Forward selection}
\usage{
fsel(mu_p, x, b_p, w, dis_p, funs, avg, d, mc.cores)
}
\arguments{
\item{mu_p}{Fitted values of the full model.}

\item{x}{A model matrix.}

\item{b_p}{Sampled estimates of the coefficient of the full model.}

\item{w}{Observation weights.}

\item{dis_p}{dispersion parameter of the full model.}

\item{funs}{List of family-specific functions for the NR.}

\item{avg}{If TRUE, KL divergence is calculated using average of the full model, rather than individual samples.}

\item{d}{Maximum number of features in the projection (incl. intercept).}

\item{mc.cores}{Number of cores used.}
}
\description{
Performs forward selection on the data and returns
the 'optimal' order of the parameters.
}

