% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glm_proj.R
\name{glm_proj}
\alias{glm_proj}
\title{GLM projection to submodel.}
\usage{
glm_proj(fit, n_out, n_sel, d, mc.cores)
}
\arguments{
\item{fit}{Either a stanreg object or a list with the model matrix, the weight vector
a \code{\link{family}}-object and the dispersion parameter, if applicable.}

\item{n_out}{Number of samples used in the projection.
If not provided, \code{min(400, ncol(b))} is used.}

\item{n_sel}{Number of samples used in the variable selection.
If not provided, average of mu is used instead of the samples.}

\item{d}{Maximum number of features in the projection (incl. intercept).
If not provided, \code{ncol(x)} - 1 is used.}

\item{mc.cores}{Number of cores used when averaging KL divergence over samples.
If not provided, defaults to \code{getOption("loo.cores", parallel::detectCores())}.}
}
\value{
A list with class \code{'glmproj'} containing
sequence of features added, distributions of the parameters
of the submodels, the corresponding KL divergences and the model family.
}
\description{
GLM projection to submodel.
}

