% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IRLS.R
\name{IRLS}
\alias{IRLS}
\title{Iteratively reweighted least squared}
\usage{
IRLS(p_full, d_train, b0, args)
}
\description{
Does the projection from the full model to the submodel.
}
\details{
Written by hand because glm.fit in R is too slow.
}

