% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IRLS.R
\name{IRLS}
\alias{IRLS}
\title{Iteratively reweighted least squared}
\usage{
IRLS(p_full, d_train, family_kl, intercept, regul = 1e-12, eps = 1e-12,
  max_it = 300)
}
\description{
Does the projection from the full model to the submodel.
}
\details{
Written by hand because glm.fit in R is too slow.
- Initialization of b currenly quite ad hoc
- rewrite in C?
}

