% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kl_helpers.R
\name{kl_gauss}
\alias{kl_gauss}
\title{Model-specific helper functions.}
\usage{
kl_gauss(p, d, q)
}
\description{
Contains functions for calculating KL divergence and its first and second
derivatives for different families and link functions.
}
\details{
The NR-algorithm is general in the sense that in order to add a new model /
link function one simply needs to implement corresponding KL-divergence,
it's first and second derivatives + first and second derivatives of the
mu-parameter w.r.t eta.
}

