% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kl_over_samples.R
\name{kl_over_samples}
\alias{kl_over_samples}
\title{KL divergence over samples}
\usage{
kl_over_samples(mu_p, x, b_p, w, dis_p, funs)
}
\arguments{
\item{\code{mu_p}}{Fitted values of the full model.}

\item{\code{x}}{A model matrix of the selected variables.}

\item{\code{b_p}}{Sampled estimates of the coefficients of the selected variables.}

\item{\code{w}}{Observation weights.}

\item{\code{dis_p}}{dispersion parameter of the full model.}

\item{\code{funs}}{List of family-specific functions for the NR.}
}
\description{
Calculates the KL divergence over ns samples.
}

