% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NR.R
\name{NR}
\alias{NR}
\title{Newton-Rhapson}
\usage{
NR(p_full, d_train, b0, family_kl, max_it = 30, eps = 1e-10)
}
\description{
Minimizes the KL divergence between the full model and the submodel
using Newton-Rhapson algorithm.
}
\details{
Naming convention described in fsel.R.
}

