% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NR.R
\name{NR}
\alias{NR}
\title{Newton-Rhapson}
\usage{
NR(p, d, b0, family, max_it = 30, eps = 1e-10)
}
\arguments{
\item{\code{mu_p}}{Fitted values of the full model.}

\item{\code{x}}{A model matrix of the selected variables.}

\item{\code{b_p}}{Sampled estimates of the coefficient of the full model.}

\item{\code{w}}{Observation weights.}

\item{\code{dis_p}}{dispersion parameter of the full model.}

\item{\code{funs}}{Model-specific helper functions.}

\item{\code{max_it}}{Maximum number of iterations for the algorithm. Defaults to 30.}

\item{\code{eps}}{Tolerance, when derivative to any direction is at most eps, algorithm stops. Defaults to 1e-10.}
}
\description{
Minimizes the KL divergence between the full model and the submodel
using Newton-Rhapson algorithm.
}

