% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{plot.vsel}
\alias{plot.vsel}
\title{Plot summary statistics of a variable selection}
\usage{
\method{plot}{vsel}(
  x,
  nterms_max = NULL,
  stats = "elpd",
  deltas = FALSE,
  alpha = 0.32,
  baseline = if (!inherits(x$refmodel, "datafit")) "ref" else "best",
  ...
)
}
\arguments{
\item{x}{An object of class \code{vsel} (returned by \code{\link[=varsel]{varsel()}} or \code{\link[=cv_varsel]{cv_varsel()}}).}

\item{nterms_max}{Maximum submodel size for which the statistics are
calculated. Note that \code{nterms_max} does not count the intercept, so use
\code{nterms_max = 0} for the intercept-only model. For \code{\link[=plot.vsel]{plot.vsel()}},
\code{nterms_max} must be at least \code{1}.}

\item{stats}{One or several strings determining which statistics to
calculate. Available statistics are:
\itemize{
\item \code{elpd}: (expected) sum of log predictive densities;
\item \code{mlpd}: mean log predictive density, that is, \code{elpd} divided by the
number of observations (data points);
\item \code{mse}: mean squared error;
\item \code{rmse}: root mean squared error;
\item \code{acc} (or its alias, \code{pctcorr}): classification accuracy (\code{\link[=binomial]{binomial()}}
family only);
\item \code{auc}: area under the ROC curve (\code{\link[=binomial]{binomial()}} family only).
}}

\item{deltas}{If \code{TRUE}, the submodel statistics are estimated relative to
the baseline model (see argument \code{baseline}) instead of estimating the
actual values of the statistics.}

\item{alpha}{A number giving the desired coverage of the credible intervals.
For example, \code{alpha = 0.32} corresponds to 68\% probability mass within the
intervals, that is, one-standard-error intervals.}

\item{baseline}{Either \code{"ref"} or \code{"best"} indicating whether the baseline is
the reference model or the best submodel found, respectively.}

\item{...}{Currently ignored.}
}
\description{
This is the \code{\link[=plot]{plot()}} method for \code{vsel} objects (returned by \code{\link[=varsel]{varsel()}} or
\code{\link[=cv_varsel]{cv_varsel()}}).
}
\examples{
if (requireNamespace("rstanarm", quietly = TRUE)) {
  # Data:
  dat_gauss <- data.frame(y = df_gaussian$y, df_gaussian$x)

  # The "stanreg" fit which will be used as the reference model (with small
  # values for `chains` and `iter`, but only for technical reasons in this
  # example; this is not recommended in general):
  fit <- rstanarm::stan_glm(
    y ~ X1 + X2 + X3 + X4 + X5, family = gaussian(), data = dat_gauss,
    QR = TRUE, chains = 2, iter = 500, refresh = 0, seed = 9876
  )

  # Variable selection (here without cross-validation and with small values
  # for `nterms_max`, `nclusters`, and `nclusters_pred`, but only for the
  # sake of speed in this example; this is not recommended in general):
  vs <- varsel(fit, nterms_max = 3, nclusters = 5, nclusters_pred = 10,
               seed = 5555)
  print(plot(vs))
}

}
