% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.R
\name{predict.glmproj}
\alias{predict.glmproj}
\title{Predict}
\usage{
\method{predict}{glmproj}(object, x, d)
}
\arguments{
\item{object}{A glmproj object.}

\item{x}{A model matrix.}

\item{d}{The number of features to be used in the prediction.
If not provided, uses all features.}
}
\description{
A predict method for glmproj objects.
}

