% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/refmodel.R
\name{predict.refmodel}
\alias{predict.refmodel}
\title{Predictions using the reference model}
\usage{
\method{predict}{refmodel}(object, xnew, ynew = NULL, offsetnew = NULL,
  weightsnew = NULL, type = "response", ...)
}
\arguments{
\item{object}{The object returned by \link{varsel},
\link{cv_varsel} or \link{project}.}

\item{xnew}{The predictor values used in the prediction. If \code{vind} is
specified, then \code{xnew} should either be a dataframe containing column names
that correspond to \code{vind} or a matrix with the number and order of columns
corresponding to \code{vind}. If \code{vind} is unspecified, then \code{xnew} must
either be a dataframe containing all the column names as in the original data or a matrix
with the same columns at the same positions as in the original data.}

\item{ynew}{New (test) target variables. If given, then the log predictive density
for the new observations is computed.}

\item{offsetnew}{Offsets for the new observations. By default a vector of
zeros.}

\item{weightsnew}{Weights for the new observations. For binomial model,
corresponds to the number trials per observation. For \code{proj_linpred},
this argument matters only if \code{ynew} is specified. By default a vector
of ones.}

\item{type}{Scale on which the predictions are returned. Either 'link' (the latent function
value, from -inf to inf) or 'response' (the scale on which the target \code{y} is measured, 
obtained by taking the inverse-link from the latent value).}

\item{...}{Additional argument passed to \link{project} if \code{object}
is an object returned by \link{varsel} or \link{cv_varsel}.}
}
\value{
An object that can be passed to all the functions that
take the reference fit as the first argument, such as \link{varsel}, \link{cv_varsel},
\link[=proj-pred]{proj_predict} and \link[=proj-pred]{proj_linpred}.
}
\description{
Compute the predictions using the reference model, that is, compute the
expected value for the next observation, or evaluate the log-predictive
density at a given point.
}
