% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/refmodel.R
\name{predict.refmodel}
\alias{predict.refmodel}
\title{Predict method for reference model objects}
\usage{
\method{predict}{refmodel}(
  object,
  newdata = NULL,
  ynew = NULL,
  offsetnew = NULL,
  weightsnew = NULL,
  type = "response",
  ...
)
}
\arguments{
\item{object}{The object of class \code{refmodel}.}

\item{newdata}{Passed to argument \code{newdata} of the reference model's
\code{extract_model_data} function (see \code{\link[=init_refmodel]{init_refmodel()}}). Provides the
predictor (and possibly also the response) data for the new (or old)
observations.}

\item{ynew}{New (test) target variables. If given, then the log predictive
density for the new observations is computed.}

\item{offsetnew}{Passed to argument \code{orhs} of the reference model's
\code{extract_model_data} function (see \code{\link[=init_refmodel]{init_refmodel()}}). Used to get the
offsets for the new (or old) observations.}

\item{weightsnew}{Passed to argument \code{wrhs} of the reference model's
\code{extract_model_data} function (see \code{\link[=init_refmodel]{init_refmodel()}}). Used to get the
weights for the new (or old) observations.}

\item{type}{Scale on which the predictions are returned. Either 'link' (the
latent function value, from -inf to inf) or 'response' (the scale on which
the target \code{y} is measured, obtained by taking the inverse-link from the
latent value).}

\item{...}{Currently ignored.}
}
\value{
Returns either a vector of predictions, or vector of log predictive
densities evaluated at \code{ynew} if \code{ynew} is not \code{NULL}.
}
\description{
Compute the predictions using the reference model, that is, compute the
expected value for the next observation, or evaluate the log-predictive
density at a given point.
}
\details{
Argument \code{weightsnew} is only relevant if \code{!is.null(ynew)}.
}
