% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{print-vsel}
\alias{print-vsel}
\alias{print.vsel}
\title{Print methods for vsel/vsel objects}
\usage{
\method{print}{vsel}(x, digits = 1, ...)
}
\arguments{
\item{x}{An object of class vsel/vsel.}

\item{digits}{Number of decimal places to be reported (1 by default).}

\item{...}{Further arguments passed to \code{\link{summary.vsel}}.}
}
\value{
Returns invisibly the data frame produced by
  \code{\link{summary.vsel}}.
}
\description{
The \code{print} methods for vsel/vsel objects created by
\code{\link{varsel}} or \code{\link{cv_varsel}}) rely on
\code{\link{summary.vsel}} to display a summary of the results of the
projection predictive variable selection.
}
