% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{print.vselsummary}
\alias{print.vselsummary}
\title{Print summary of variable selection}
\usage{
\method{print}{vselsummary}(x, ...)
}
\arguments{
\item{x}{An object of class \code{vselsummary}.}

\item{...}{Passed to \code{\link[=print.data.frame]{print.data.frame()}}.}
}
\value{
The output of \code{\link[=summary.vsel]{summary.vsel()}} (invisible).
}
\description{
This is the \code{\link[=print]{print()}} method for summary objects created by \code{\link[=summary.vsel]{summary.vsel()}}.
It displays a summary of the results of the projection predictive variable
selection.
}
