% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/project.R
\name{project}
\alias{project}
\title{Projection to a submodels of selected size.}
\usage{
project(object, nv = NULL, ns = NULL, nc = NULL, intercept = NULL,
  seed = NULL, ...)
}
\arguments{
\item{object}{The object returned by \link[=varsel]{varsel} or
\link[=cv_varsel]{cv_varsel}.}

\item{nv}{Number of variables in the submodel. Can also be a list, in
which the projection is performed for each of the sizes in the list. If
\code{NULL}, the projection is performed for all possible submodel sizes.}

\item{ns}{Same as in \link[=varsel]{varsel}.}

\item{nc}{Same as in \link[=varsel]{varsel}.}

\item{intercept}{Same as in \link[=varsel]{varsel}.}

\item{seed}{A seed used in the clustering (if \code{nc!=ns}). Can be used
to ensure same results every time.}

\item{...}{Currently ignored.}
}
\value{
The original fit-object object augmented with a field 'proj',
which is a list containing the following elements:
\describe{
 \item{\code{p_sub}}{A list of the projected parameters for each of the
 submodels.}
 \item{\code{intercept}}{A logical indicating whether the submodels contain
  intercept or not.}
}
}
\description{
Perform the projection predictive variable selection for a generalized
linear model fitted with rstanarm.
}
\examples{
\dontrun{
### Usage with stanreg objects
fit <- stan_glm(y~x, binomial())
fit_v <- varsel(fit)
fit_p <- project(fit_v)
}

}
