% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/project.R
\name{project}
\alias{project}
\title{Projection to submodels of selected sizes}
\usage{
project(object, nv = NULL, vind = NULL, ns = NULL, nc = NULL,
  intercept = NULL, seed = NULL, return_fit = TRUE, ...)
}
\arguments{
\item{object}{The object returned by \link[=varsel]{varsel} or
\link[=cv_varsel]{cv_varsel}.}

\item{nv}{Number of variables in the submodel (the variable combination is taken from the
\code{varsel} information). If a list, then the projection is performed for each model size.
Default is all model sizes up to the maximum number of variables in \code{varsel}.
 Ignored if \code{vind} is specified.}

\item{vind}{Variable indices onto which the projection is done. If specified, \code{nv} is ignored.}

\item{ns}{Number of samples to be projected. Ignored if \code{nc} is specified.}

\item{nc}{Number of clusters in the clustered projection. Default is 50.}

\item{intercept}{Whether to use intercept. Default is \code{TRUE}.}

\item{seed}{A seed used in the clustering (if \code{nc!=ns}). Can be used
to ensure same results every time.}

\item{...}{Currently ignored.}
}
\value{
The original fit-object object augmented with a field 'proj',
which is a list of submodels (or a single submodel if projection was performed onto
a single variable combination).
}
\description{
Perform the projection predictive variable selection for a generalized
linear model fitted with rstanarm.
}
\examples{
\dontrun{
### Usage with stanreg objects
fit <- stan_glm(y~x, binomial())
fit_v <- varsel(fit)
fit_p <- project(fit_v)
}

}
