% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/project.R
\name{project}
\alias{project}
\title{Projection onto submodel(s)}
\usage{
project(
  object,
  nterms = NULL,
  solution_terms = NULL,
  cv_search = TRUE,
  ndraws = 400,
  nclusters = NULL,
  seed = NULL,
  regul = 1e-04,
  ...
)
}
\arguments{
\item{object}{Either a \code{refmodel}-type object created by
\link[=get_refmodel]{get_refmodel} or \link[=init_refmodel]{init_refmodel},
or an object which can be converted to a reference model using
\link[=get_refmodel]{get_refmodel}.}

\item{nterms}{Number of terms in the submodel (the variable combination is
taken from the \code{varsel} information). If a numeric vector, then the
projection is performed for each model size. If \code{NULL}, the model size
suggested by the variable selection (see function \code{suggest_size}).
Ignored if \code{solution_terms} is specified. Note that \code{nterms} does
not count the intercept, so use \code{nterms = 0} for the intercept-only
model.}

\item{solution_terms}{Variable indices onto which the projection is done. If
specified, \code{nterms} is ignored.}

\item{cv_search}{If TRUE, then the projected coefficients after L1-selection
are computed without any penalization (or using only the regularization
determined by \code{regul}). If FALSE, then the coefficients are the
solution from the L1-penalized projection. This option is relevant only if
L1-search was used. Default is TRUE for genuine reference models and FALSE
if \code{object} is datafit (see \link[=init_refmodel]{init_refmodel}).}

\item{ndraws}{Number of posterior draws to be projected. Cannot be larger
than the number of draws in the reference model. \strong{Caution:} For
\code{ndraws <= 20}, the value of \code{ndraws} is passed to
\code{nclusters} (so that clustering is used). Ignored if \code{nclusters}
is not \code{NULL} or if the reference model is of class \code{"datafit"}
(in which case one cluster is used). See also section "Details" below.}

\item{nclusters}{Number of clusters of posterior draws to be projected.
Ignored if the reference model is of class \code{"datafit"} (in which case
one cluster is used). For the meaning of \code{NULL}, see argument
\code{ndraws}. See also section "Details" below.}

\item{seed}{A seed used for clustering the reference model's posterior draws
(if \code{!is.null(nclusters)}). Can be used to ensure reproducible
results. If \code{NULL}, no seed is set and therefore, the results are not
reproducible. See \code{\link{set.seed}} for details.}

\item{regul}{Amount of ridge regularization when fitting the models in the
projection. Usually there is no need for regularization, but sometimes for
some models the projection can be ill-behaved and we need to add some
regularization to avoid numerical problems.}

\item{...}{Arguments passed to \link[=get_refmodel]{get_refmodel}.}
}
\value{
If the projection is performed onto a single submodel (i.e.,
\code{nterms} has length one or \code{solution_terms} is specified), an
object of class \code{"projection"} which is a \code{list} containing the
following elements:
\describe{
\item{\code{dis}}{Projected draws for the dispersion parameter.}
\item{\code{kl}}{The KL divergence from the submodel to the reference
model.}
\item{\code{weights}}{Weights for the projected draws.}
\item{\code{solution_terms}}{A character vector of the submodel's
predictor terms, ordered the way in which the terms were added to the
submodel.}
\item{\code{sub_fit}}{The submodel's fitted model object.}
\item{\code{family}}{A modified \code{\link{family}}-object.}
\item{\code{p_type}}{A single logical value indicating whether the
reference model's posterior draws have been clustered for the projection
(\code{TRUE}) or not (\code{FALSE}).}
\item{\code{intercept}}{A single logical value indicating whether the
reference model (as well as the submodel) contains an intercept
(\code{TRUE}) or not (\code{FALSE}).}
\item{\code{extract_model_data}}{The \code{extract_model_data()} function
from the reference model (see \code{\link{init_refmodel}}).}
\item{\code{refmodel}}{The reference model object (see
\code{\link{init_refmodel}}).}
}
If the projection is performed onto more than one submodel, the output from
above is returned for each submodel, giving a \code{list} with one element
for each submodel.
}
\description{
Project the reference model onto a single submodel consisting of a specific
combination of predictor terms or onto a single or multiple submodels of
specific sizes.
}
\details{
Using less draws or clusters in \code{ndraws} or \code{nclusters}
than posterior draws in the reference model may result in slightly
inaccurate projection performance. Increasing these arguments linearly
affects the computation time.
}
\examples{
\donttest{
if (requireNamespace("rstanarm", quietly = TRUE)) {
  ### Usage with stanreg objects
  n <- 30
  d <- 5
  x <- matrix(rnorm(n * d), nrow = n)
  y <- x[, 1] + 0.5 * rnorm(n)
  data <- data.frame(x, y)

  fit <- rstanarm::stan_glm(y ~ X1 + X2 + X3 + X4 + X5, gaussian(),
    data = data, chains = 2, iter = 500)
  vs <- varsel(fit)

  # project onto the best model with 4 variables
  proj4 <- project(vs, nterms = 4)

  # project onto an arbitrary variable combination
  proj <- project(fit, solution_terms = c("X1", "X3", "X5"))
}
}

}
