% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/projpred-package.R
\docType{package}
\name{projpred-package}
\alias{projpred}
\alias{projpred-package}
\title{Projection predictive feature selection}
\description{
\pkg{projpred} is an \R package for performing a projection predictive
variable (or "feature") selection for generalized linear models (GLMs),
generalized linear multilevel (or "mixed") models (GLMMs), generalized
additive models (GAMs), and generalized additive multilevel (or "mixed")
models (GAMMs). Note that the term "generalized" includes the Gaussian family
as well.

The package is compatible with \pkg{rstanarm} and \pkg{brms}, but developers
of other packages are welcome to add more methods for the \code{\link[=get_refmodel]{get_refmodel()}}
generic. Custom reference models can also be used via \code{\link[=init_refmodel]{init_refmodel()}}.

Currently, the supported families are \code{\link[=gaussian]{gaussian()}}, \code{\link[=binomial]{binomial()}} (and, via
\code{\link[brms:get_refmodel.brmsfit]{brms::get_refmodel.brmsfit()}}, \code{\link[brms:brmsfamily]{brms::bernoulli()}}), and \code{\link[=poisson]{poisson()}}.

See the vignettes
(\href{https://mc-stan.org/projpred/articles/quickstart.html}{quickstart-vignette}
and
\href{https://mc-stan.org/projpred/articles/quickstart_glmm.html}{quickstart-glmm-vignette})
for example applications. Smaller examples are included here in the
documentation.
}
\section{Functions}{
\describe{
\item{\code{\link[=varsel]{varsel()}}, \code{\link[=cv_varsel]{cv_varsel()}}}{Perform the variable selection, possibly
with cross-validation (CV).}
\item{\code{\link[=summary.vsel]{summary.vsel()}}, \code{\link[=print.vsel]{print.vsel()}}, \code{\link[=plot.vsel]{plot.vsel()}},
\code{\link[=suggest_size.vsel]{suggest_size.vsel()}}, \code{\link[=solution_terms.vsel]{solution_terms.vsel()}}}{Post-process the results
from the variable selection.}
\item{\code{\link[=project]{project()}}}{Project the reference model onto submodel(s). Typically,
this follows the variable selection, but it can also be applied directly
(without a variable selection).}
\item{\code{\link[=as.matrix.projection]{as.matrix.projection()}}}{Extract projected parameter draws.}
\item{\code{\link[=proj_linpred]{proj_linpred()}}, \code{\link[=proj_predict]{proj_predict()}}}{Make predictions from a submodel
(after projecting the reference model onto it).}
}
}

\section{References}{
See the \code{CITATION} file for details on citation.
\itemize{
\item Goutis, C. and Robert, C. P. (1998). Model choice in generalised linear
models: A Bayesian approach via Kullback–Leibler projections. \emph{Biometrika},
\strong{85}(1):29–37.
\item Dupuis, J. A. and Robert, C. P. (2003). Variable selection in qualitative
models via an entropic explanatory power. \emph{Journal of Statistical Planning
and Inference}, \strong{111}(1-2):77–94. DOI:
\href{https://doi.org/10.1016/S0378-3758(02)00286-0}{10.1016/S0378-3758(02)00286-0}.
\item Piironen, J. and Vehtari, A. (2017). Comparison of Bayesian predictive
methods for model selection. \emph{Statistics and Computing}, \strong{27}(3):711-735.
DOI: \href{https://doi.org/10.1007/s11222-016-9649-y}{10.1007/s11222-016-9649-y}.
\item Piironen, J., Paasiniemi, M., and Vehtari, A. (2020). Projective inference
in high-dimensional problems: Prediction and feature selection. \emph{Electronic
Journal of Statistics}, \strong{14}(1):2155-2197. DOI:
\href{https://doi.org/10.1214/20-EJS1711}{10.1214/20-EJS1711}.
\item Catalina, A., Bürkner, P.-C., and Vehtari, A. (2020). Projection predictive
inference for generalized linear and additive multilevel models.
\emph{arxiv:2010.06994}. URL: \url{https://arxiv.org/abs/2010.06994}.
}
}

\seealso{
Useful links:
\itemize{
  \item \url{https://mc-stan.org/projpred/}
  \item \url{https://discourse.mc-stan.org}
  \item Report bugs at \url{https://github.com/stan-dev/projpred/issues/}
}

}
\author{
\strong{Maintainer}: Alejandro Catalina \email{alecatfel@gmail.com}

Authors:
\itemize{
  \item Juho Piironen \email{juho.t.piironen@gmail.com}
  \item Markus Paasiniemi
  \item Aki Vehtari
}

Other contributors:
\itemize{
  \item Jonah Gabry [contributor]
  \item Marco Colombo [contributor]
  \item Paul-Christian Bürkner [contributor]
  \item Hamada S. Badr [contributor]
  \item Frank Weber [contributor]
}

}
