% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{solution_terms}
\alias{solution_terms}
\title{Recovers solution path from a variable selection object.}
\usage{
solution_terms(object)
}
\arguments{
\item{object}{A vsel object returned by \link[=varsel]{varsel} or
\link[=cv_varsel]{cv_varsel}.}
}
\value{
Variable selection solution path
}
\description{
Recovers solution path from a variable selection object.
}
