% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{varsel-methods}
\alias{varsel-methods}
\alias{proj_linpred}
\alias{proj_coef}
\alias{proj_se}
\alias{proj_sigma}
\title{Methods for objects with varsel-information}
\usage{
proj_linpred(object, xnew, ynew = NULL, offsetnew = NULL,
  weightsnew = NULL, transform = FALSE, integrated = FALSE, nv = NULL,
  vind = NULL, ns = NULL, nc = NULL, ...)

proj_coef(object, ...)

proj_se(object, ...)

proj_sigma(object, ...)
}
\arguments{
\item{object}{The object returned by \link[=varsel]{varsel} or
\link[=cv_varsel]{cv_varsel}.}

\item{xnew}{The predictor values used in the prediction. The number and order of the columns
should be the same as in the original full data if argument \code{nv} is specified (see below).
However, if argument \code{vind} is specified, then the number and order of columns should
correspond to \code{vind}.}

\item{ynew}{New (test) target variables.}

\item{offsetnew}{Offsets for the new observations.}

\item{weightsnew}{Weights for the new observations. This argument matters only if \code{ynew} is specified.}

\item{transform}{Should the linear predictor be transformed using the inverse-link function?
Default is \code{FALSE}.}

\item{integrated}{If \code{TRUE}, the output is averaged over the
parameters. Default is \code{FALSE}.}

\item{nv}{Number of variables in the submodel (the variable combination is taken from the
\code{varsel} information). If a list, then results for all specified
model sizes are returned. Ignored if \code{vind} is specified.}

\item{vind}{Variable indices with which to predict. If specified, \code{nv} is ignored.}

\item{ns}{Number of samples to be projected. Ignored if \code{nc} is specified.}

\item{nc}{Number of clusters in the clustered projection. Default is 50.}

\item{...}{Currently ignored.}

\item{intercept}{Whether to use intercept. Default is \code{TRUE}.}
}
\description{
The methods can be used to extract all the information related to variable
selection obtained running either \link[=varsel]{varsel} or
\link[=cv_varsel]{cv_varsel}. Most of the functions work similar to the
generic methods defined for objects of the class 'lm', 'glm', 'stanreg',
eg. proj_coef corresponds to \link[=coef]{coef}, proj_se corresponds to
\link[=se]{se} etc. Note that these are under development and subject might
change in the future.
}
