% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{varsel-methods}
\alias{varsel-methods}
\alias{proj_linpred}
\alias{proj_coef}
\alias{proj_se}
\alias{proj_sigma}
\alias{varsel_plot}
\alias{varsel_statistics}
\title{Methods for objects with varsel-information.}
\usage{
proj_linpred(object, transform = FALSE, xnew = NULL, ynew = NULL,
  offsetnew = NULL, newdata = NULL, nv = NULL, integrated = FALSE,
  ns = NULL, nc = NULL, ...)

proj_coef(object, ...)

proj_se(object, ...)

proj_sigma(object, ...)

varsel_plot(object, ..., nv_max = NULL, statistics = NULL, deltas = T,
  n_boot = 1000, alpha = 0.1)

varsel_statistics(object, ..., nv_max = NULL, deltas = F)
}
\arguments{
\item{object}{The object returned by \link[=varsel]{varsel} or
\link[=cv_varsel]{cv_varsel}.}

\item{transform}{Same as in \link[=posterior_linpred]{posterior_linpred}.}

\item{xnew}{New data to be used with the prediction. If \code{NULL}, the
data used when fitting the model is used.}

\item{ynew}{Same as \code{xnew}, but for the ouput.}

\item{offsetnew}{Same as \code{xnew}, but for the offset.}

\item{nv}{Number of variables in the submodel. Can also be a list, in
which case a list is returned for each submodel size.}

\item{integrated}{If \code{TRUE}, the output is averaged over the
parameters. Defaults to \code{FALSE}.}

\item{ns}{Same as in \link[=varsel]{varsel}.}

\item{nc}{Same as in \link[=varsel]{varsel}.}

\item{...}{Currently ignored.}

\item{nv_max}{Maximum submodel size for which the statistics are calculated.}

\item{statistics}{A list of strings of statistics to calculate. Available
options are: kl, mse, mlpd, kl, (gaussian only), pctcorr (binomial only).
If \code{NULL}, all statistics are calculated.}

\item{deltas}{If \code{TRUE}, the difference between the full model and the
submodel is returned instead of the actual value of the statistic.
Defaults to \code{FALSE}.}

\item{n_boot}{Number of bootstrap samples for calculating the credible
intervals of the statistics.}

\item{alpha}{A number indicating the desired coverage of the credible
intervals. Eg. \code{alpha=0.1} corresponds to 90% probability mass
within the intervals. Defaults to \code{0.1}.}
}
\description{
The methods can be used to extract all the information related to variable
selection obtained running either \link[=varsel]{varsel} or
\link[=cv_varsel]{cv_varsel}. Most of the functions work similar to the
generic methods defined for objects of the class 'lm', 'glm', 'stanreg',
eg. proj_coef corresponds to \link[=coef]{coef}, proj_linpred corresponds to
\link[=se]{se} etc. Note that these are under development and subject might
change in the future.
}
