% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{varsel-statistics}
\alias{varsel-statistics}
\alias{varsel_plot}
\alias{varsel_statistics}
\title{Plotting or printing summary statistics related to variable selection}
\usage{
varsel_plot(object, ..., nv_max = NULL, statistics = NULL, deltas = T,
  n_boot = 1000, alpha = 0.1)

varsel_statistics(object, ..., nv_max = NULL, deltas = F)
}
\arguments{
\item{object}{The object returned by \link[=varsel]{varsel} or
\link[=cv_varsel]{cv_varsel}.}

\item{nv_max}{Maximum submodel size for which the statistics are calculated.}

\item{statistics}{A list of strings of statistics to calculate. Available
options are: kl, mse, mlpd, kl, (gaussian only), pctcorr (binomial only).
If \code{NULL}, all statistics are calculated.}

\item{deltas}{If \code{TRUE}, the difference between the full model and the
submodel is returned instead of the actual value of the statistic.
Defaults to \code{FALSE}.}

\item{n_boot}{Number of bootstrap samples for calculating the credible
intervals of the statistics.}

\item{alpha}{A number indicating the desired coverage of the credible
intervals. Eg. \code{alpha=0.1} corresponds to 90% probability mass
within the intervals. Defaults to \code{0.1}.}
}
\description{
\code{varsel_statistics} can be used to obtain summary statistics related to
variable selection. The same statistics can be plotted with
\code{varsel_plot}.
}
