% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/varsel.R
\name{varsel}
\alias{varsel}
\alias{varsel.default}
\alias{varsel.refmodel}
\title{Variable selection (without cross-validation)}
\usage{
varsel(object, ...)

\method{varsel}{default}(object, ...)

\method{varsel}{refmodel}(
  object,
  d_test = NULL,
  method = NULL,
  ndraws = 20,
  nclusters = NULL,
  ndraws_pred = 400,
  nclusters_pred = NULL,
  cv_search = !inherits(object, "datafit"),
  nterms_max = NULL,
  verbose = TRUE,
  lambda_min_ratio = 1e-05,
  nlambda = 150,
  thresh = 1e-06,
  regul = 1e-04,
  penalty = NULL,
  search_terms = NULL,
  seed = NULL,
  ...
)
}
\arguments{
\item{object}{Either a \code{refmodel}-type object created by \code{\link[=init_refmodel]{init_refmodel()}},
an object which can be converted to a reference model using
\code{\link[=get_refmodel]{get_refmodel()}}, or a \code{vsel} object resulting from \code{\link[=varsel]{varsel()}} or
\code{\link[=cv_varsel]{cv_varsel()}}.}

\item{...}{Additional arguments to be passed to the \code{\link[=get_refmodel]{get_refmodel()}}
function.}

\item{d_test}{For internal use only. A \code{list} providing information about the
test set which is used to evaluate model performance. If not provided, the
training set is used.}

\item{method}{The search method, i.e. the method for finding a single
submodel per number of terms. Possible options are \code{"L1"} for L1 search and
\code{"forward"} for forward search. Default is \code{"forward"} if the reference
model has multilevel or additive terms and \code{"L1"} otherwise.}

\item{ndraws}{Number of posterior draws used in the variable selection.
Automatically truncated at the number of draws in the reference model
(which is \code{1} for \code{datafit}s). \strong{Caution:} For \code{ndraws <= 20}, the value
of \code{ndraws} is passed to \code{nclusters} (so that clustering is used). Ignored
if \code{nclusters} is not \code{NULL} or if \code{method == "L1"} (L1 search uses always
one cluster). See also section "Details" below.}

\item{nclusters}{Number of clusters of posterior draws used in the variable
selection. Ignored if \code{method == "L1"} (L1 search uses always one cluster).
For the meaning of \code{NULL}, see argument \code{ndraws}. See also section
"Details" below.}

\item{ndraws_pred}{Number of posterior draws used for prediction (after
selection). Automatically truncated at the number of draws in the reference
model (which is \code{1} for \code{datafit}s). \strong{Caution:} For \code{ndraws_pred <= 20},
the value of \code{ndraws_pred} is passed to \code{nclusters_pred} (so that
clustering is used). Ignored if \code{nclusters_pred} is not \code{NULL}. See also
section "Details" below.}

\item{nclusters_pred}{Number of clusters of posterior draws used for
prediction (after selection). For the meaning of \code{NULL}, see argument
\code{ndraws_pred}. See also section "Details" below.}

\item{cv_search}{If \code{TRUE}, then the projected coefficients after an L1
search are computed without any penalization (or using only the
regularization determined by \code{regul}). If \code{FALSE}, then the coefficients
are the solution from the L1-penalized projection. This option is relevant
only if \code{method == "L1"}. Default is \code{TRUE} for genuine reference models
and \code{FALSE} if \code{object} is datafit (see \code{\link[=init_refmodel]{init_refmodel()}}).}

\item{nterms_max}{Maximum number of variables until which the selection is
continued. Defaults to \code{min(19, D)} where \code{D} is the number of terms in the
reference model (or in \code{search_terms}, if supplied). Note that \code{nterms_max}
does not count the intercept, so use \code{nterms_max = 0} for the
intercept-only model.}

\item{verbose}{A single logical value indicating whether to print out
additional information while running (\code{TRUE}) or not (\code{FALSE}).}

\item{lambda_min_ratio}{Ratio between the smallest and largest lambda in the
L1-penalized search. This parameter essentially determines how long the
search is carried out, i.e., how large submodels are explored. No need to
change the default value unless the program gives a warning about this.}

\item{nlambda}{Number of values in the lambda grid for L1-penalized search.
No need to change unless the program gives a warning about this.}

\item{thresh}{Convergence threshold when computing the L1 path. Usually,
there is no need to change this.}

\item{regul}{A number giving the amount of ridge regularization when
projecting onto (i.e., fitting) submodels which are (G)LMs. Usually there
is no need for regularization, but sometimes we need to add (or rather
increase, given that \code{regul} defaults to \code{1e-4}) some regularization to
avoid numerical problems.}

\item{penalty}{Vector determining the relative penalties or costs for the
variables. A value of \code{0} means that those variables have no cost and will
therefore be selected first, whereas \code{Inf} means those variables will never
be selected. Currently works only if \code{method == "L1"}. By default \code{1} for
each variable.}

\item{search_terms}{A custom character vector of terms to consider for
selection. The intercept (\code{"1"}) needs to be included explicitly. The
default considers all the terms in the reference model's formula.}

\item{seed}{Pseudorandom number generation (PRNG) seed by which the same
results can be obtained again if needed. If \code{NULL}, no seed is set and
therefore, the results are not reproducible. See \code{\link[=set.seed]{set.seed()}} for details.
Here, this seed is used for clustering the reference model's posterior
draws (if \code{!is.null(nclusters)}).}
}
\value{
An object of class \code{vsel}. The elements of this object are not meant
to be accessed directly but instead via helper functions (see the vignettes
or type \code{?projpred}).
}
\description{
Perform the projection predictive variable selection for (G)LMs, (G)LMMs,
(G)AMs, and (G)AMMs.
}
\details{
Notes:
\itemize{
\item Using less draws or clusters in \code{ndraws}, \code{nclusters}, \code{nclusters_pred}, or
\code{ndraws_pred} than posterior draws in the reference model may result in
slightly inaccurate projection performance. Increasing these arguments
affects the computation time linearly.
\item An L1 search may select interaction terms before the corresponding main
terms are selected. If this is undesired, choose the forward search instead.
}
}
\examples{
if (requireNamespace("rstanarm", quietly = TRUE)) {
  # Data:
  dat_gauss <- data.frame(y = df_gaussian$y, df_gaussian$x)

  # The "stanreg" fit which will be used as the reference model:
  fit <- rstanarm::stan_glm(
    y ~ X1 + X2 + X3 + X4 + X5, family = gaussian(), data = dat_gauss,
    QR = TRUE, chains = 2, iter = 500, refresh = 0, seed = 9876
  )

  # Variable selection (here without cross-validation and with small values
  # for `nterms_max`, `nclusters`, and `nclusters_pred`, but only for the
  # sake of speed in this example; this is not recommended in general):
  vs <- varsel(fit, nterms_max = 3, nclusters = 5, nclusters_pred = 10,
               seed = 5555)
  # Now see, for example, `?print.vsel`, `?plot.vsel`, `?suggest_size.vsel`,
  # and `?solution_terms.vsel` for possible post-processing functions.
}

}
\seealso{
\code{\link[=cv_varsel]{cv_varsel()}}
}
