% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/varsel.R
\name{varsel}
\alias{varsel}
\title{Variable selection for generalized linear models}
\usage{
varsel(fit, d_test = NA, method = "L1", ...)
}
\arguments{
\item{fit}{A \link[=stanreg-objects]{stanreg} object.}

\item{d_test}{A test dataset, which is used to evaluate model performance.
If not provided, training data is used.}

\item{method}{The method used in the variable selection. Possible options are
\code{'L1'} for L1-search and \code{'forward'} for forward selection.
Defaults to \code{'L1'}.}

\item{...}{Optional arguments. Possible arguments and their defaults are:
\describe{
 \item{\code{ns = min(400, [number of draws])}}{
   Number of draws used in the variable selection.
   Cannot be larger than the number of draws in the full model.}
 \item{\code{nc = 0}}{
   If nonzero, a clustering with \code{nc} clusters is performed for
   the draws and the cluster centers are used in the variable selection
   instead of the actual draws.}
 \item{\code{nv = min(ncol(x) - 1, rankMatrix(x))}}{
   Maximum number of variables to be used in the projection (incl. intercept).
   Cannot be larger than \code{min(ncol(x) - 1, rankMatrix(x))}.}
 \item{\code{verbose = FALSE}}{
   If \code{verbose = TRUE}, prints information about the progress of the
   variable selection.}
}}
}
\value{
The original \link[=stanreg-objects]{stanreg} object augmented with an element 'varsel',
which is a list containing the following elements:
\describe{
 \item{\code{chosen}}{The order in which the variables were added to the submodel.}
 \item{\code{stats}}{An array with statistics of the submodel performance.}
 \item{\code{family}}{A \code{\link{family}}-object.}
}
}
\description{
Perform the projection predictive variable selection for a generalized
linear model fitted with rstanarm.
}
\examples{
\dontrun{
### Usage with stanreg objects
fit <- stan_glm(y~x, binomial())
fit_v <- varsel(fit)
plot_varsel(fit_v)
}

}

