% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/varsel.R
\name{varsel}
\alias{varsel}
\title{Variable selection for generalized linear models}
\usage{
varsel(fit, d_test = NA, ...)
}
\arguments{
\item{fit}{A \link[=stanreg-objects]{stanreg} object.}

\item{d_test}{A test dataset, which is used to evaluate model performance. If
not provided, training data is used.}

\item{...}{Optional arguments. Possible arguments and their defaults are:
\describe{
 \item{\code{ns = min(400, [number of samples])}}{
   Number of samples used in the variable selection.
   Cannot be larger than the number of samples.}
 \item{\code{nc = 0}}{
   If nonzero, samples are clustered and the cluster centers are
   used in the variable selection instead of the actual samples.
 }
 \item{\code{nv = min(ncol(x) - 1, rankMatrix(x))}}{
   Maximum number of features to be used in the projection (incl. intercept).
   Cannot be larger than \code{min(ncol(x) - 1, rankMatrix(x))}.}
 \item{\code{verbose = FALSE}}{
   If \code{verbose = TRUE}, prints information about the progress of the
   variable selection.}
}}
}
\value{
A list with class \code{'varsel'} containing the following elements:
\describe{
 \item{\code{chosen}}{The order in which the features were added to the submodel.}
 \item{\code{stats}}{An array with statistics of the submodel performance.}
 \item{\code{family}}{A \code{\link{family}}-object.}
}
}
\description{
Perform the projection predictive variable selection for a generalized
linear model fitted with rstanarm.
}
\examples{
\dontrun{
### Usage with stanreg objects
fit <- stan_glm(y~x, binomial())
vars <- varsel(fit)
plot(vars)
}

}

